Bivariate Functions

Base.merge

PlaceholderLikelihood.add_biv_profiles_rows!

PlaceholderLikelihood.bivariate_confidenceprofile

PlaceholderLikelihood.bivariate_confidenceprofile_continuation

PlaceholderLikelihood.bivariateΨ_gradient!

PlaceholderLikelihood.boundary_smoother!

PlaceholderLikelihood.continuation_line_search!

PlaceholderLikelihood.edge_length

PlaceholderLikelihood.findNpointpairs_radialMLE!

PlaceholderLikelihood.findNpointpairs_radialrandom!

PlaceholderLikelihood.findNpointpairs_simultaneous!

PlaceholderLikelihood.get_bivariate_opt_func

PlaceholderLikelihood.get_λs_bivariate_ellipse_analytical!

PlaceholderLikelihood.heapupdates_failure!

PlaceholderLikelihood.initial_continuation_solution!

PlaceholderLikelihood.internal_angle_from_pi!

PlaceholderLikelihood.minimum_perimeter_polygon!

PlaceholderLikelihood.newboundarypoint!

PlaceholderLikelihood.refine_search_directions!

PlaceholderLikelihood.set_biv_profiles_row!

PlaceholderLikelihood.star_obj

Initialisation and Array Mapping

Missing docstring.

Missing docstring for PlaceholderLikelihood.variablemapping2dranges. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.variablemapping2d!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.boundsmapping2d!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.init_bivariate_parameters. Check Documenter's build log for details.

Likelihood Optimisation

Missing docstring.

Missing docstring for PlaceholderLikelihood.bivariateΨ!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.bivariateΨ_vectorsearch!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.bivariateΨ_continuation!. Check Documenter's build log for details.

PlaceholderLikelihood.bivariateΨ_gradient! — Function

Requires optimal values of nuisance parameters at point Ψ to be contained in p.λ_opt

Missing docstring.

Missing docstring for PlaceholderLikelihood.bivariateΨ_ellipse_analytical. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.bivariateΨ_ellipse_analytical_vectorsearch. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.bivariateΨ_ellipse_analytical_continuation. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.bivariateΨ_ellipse_analytical_gradient. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.bivariateΨ_ellipse_unbounded. Check Documenter's build log for details.

Finding Points on 2D bounds

Missing docstring.

Missing docstring for PlaceholderLikelihood.findpointonbounds. Check Documenter's build log for details.

Main Confidence Boundary Logic

Note: AnalyticalEllipseMethod is calculated using generate_N_clustered_points from EllipseSampling.jl within PlaceholderLikelihood.bivariate_confidenceprofile.

PlaceholderLikelihood.add_biv_profiles_rows! — Function

add_biv_profiles_rows!(model::LikelihoodModel, num_rows_to_add::Int)

Adds num_rows_to_add rows to model.biv_profiles_df.

PlaceholderLikelihood.set_biv_profiles_row! — Function

set_biv_profiles_row!(model::LikelihoodModel, 
    row_ind::Int, 
    θcombination::Tuple{Int, Int},
    not_evaluated_internal_points::Bool, 
    not_evaluated_predictions::Bool,
    confidence_level::Float64, 
    profile_type::AbstractProfileType,
    method::AbstractBivariateMethod, 
    num_points::Int)

Sets the columns of row row_ind of model.biv_profiles_df to contain the relevant info about a just conducted profile. model.biv_profiles_dict contains the profile for row row_ind at key row_ind.

PlaceholderLikelihood.get_bivariate_opt_func — Function

get_bivariate_opt_func(profile_type::AbstractProfileType, method::AbstractBivariateMethod)

Returns the correct bivariate optimisation function used to find the optimal values of nuisance parameters at a set of interest parameters for the profile_type log-likelihood function. The optimisation function returns the value of the profile_type log-likelihood function as well as finding the optimal nuisance parameters and saving these in one of it's inputs.

PlaceholderLikelihood.get_λs_bivariate_ellipse_analytical! — Function

get_λs_bivariate_ellipse_analytical!(boundary,
    num_points::Int,
    consistent::NamedTuple, 
    ind1::Int, 
    ind2::Int, 
    num_pars::Int,
    initGuess::Vector{<:Float64}, 
    θranges::Tuple{T, T, T}, 
    λranges::Tuple{T, T, T},
    samples_all_pars::Union{Missing, Matrix{Float64}}=missing) where T<:UnitRange

Determines the nuisance parameters for a EllipseApproxAnalytical boundary profile by optimising over the unbounded ellipse approximation of the log-likelihood centred at the MLE. At higher confidence levels, where the ellipse approximation is less accurate, it is likely that predictions produced by running the model with these optimised nuisance parameters will be unrealistic and/or the parameters themselves may be infeasible for the model.

PlaceholderLikelihood.bivariate_confidenceprofile — Function

bivariate_confidenceprofile(bivariate_optimiser::Function,
    model::LikelihoodModel, 
    num_points::Int,
    confidence_level::Float64,
    consistent::NamedTuple,
    ind1::Int,
    ind2::Int,
    profile_type::AbstractProfileType,
    method::AbstractBivariateMethod,
    mle_targetll::Float64,
    save_internal_points::Bool)

Returns a BivariateConfidenceStruct containing the num_points boundary points and internal points (if save_internal_points=true) for the specified combination of parameters ind1 and ind2, and profile_type at confidence_level using method. Calls the desired method. Called by bivariate_confidenceprofiles!.

Base.merge — Method

Base.merge(a::BivariateConfidenceStruct, b::BivariateConfidenceStruct)

Specifies how to merge two variables with type BivariateConfidenceStruct.

Minimum Perimeter Polygon

PlaceholderLikelihood.minimum_perimeter_polygon! — Function

minimum_perimeter_polygon!(points::Array{<:Real,2})

Given a set of N points that define a boundary polygon in a 2 row, N column array, solve a minimum perimeter polygon TSP problem, reorder these points in place and return the path used (vertices in order of visitation). Uses TravelingSalesmanHeuristics.jl.

Iterative Boundary Method

For IterativeBoundaryMethod.

PlaceholderLikelihood.findNpointpairs_radialMLE! — Function

Distorts uniformly spaced angles on a circle to angles on an ellipse representative of the relative magnitude of each parameter. If the magnitude of a parameter is a NaN value (i.e. either bound is Inf), then the relative magnitude is set to 1.0, as no information is known about its magnitude.

Angles are anticlockwise

PlaceholderLikelihood.edge_length — Function

edge_length(boundary, inds1, inds2, relative_magnitude)

Euclidean distance between two vertices (length of an edge), scaled by the relative magnitude of parameters, so that each dimension has roughly the same weight.

PlaceholderLikelihood.internal_angle_from_pi! — Function

internal_angle_from_pi!(vertex_internal_angle_objs, 
    indexes::UnitRange, 
    boundary, 
    adjacent_vertices)

The magnitude the internal angle in radians between two adjacent edges is from pi radians - i.e. how far away the two edges are from representing a straight boundary. If a boundary is straight then the objective is 0.0 radians, whereas if the boundary has an internal angle of pi/4 radians (45 deg) the objective is pi*3/4 (135 deg). Computes this by considering the angle between the two vectors that can be used to represent the edges (using AngleBetweenVectors.jl).

Missing docstring.

Missing docstring for PlaceholderLikelihood.internal_angle_from_pi. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.iterativeboundary_init. Check Documenter's build log for details.

PlaceholderLikelihood.newboundarypoint! — Function

newboundarypoint!(p::NamedTuple, 
    point_is_on_bounds::BitVector, 
    edge_anti_on_bounds::BitVector, 
    boundary::Matrix{Float64}, 
    boundary_all::Matrix{Float64}, 
    internal_all::Matrix{Float64}, 
    ll_values::Vector{Float64}, 
    internal_count::Int,
    bivariate_optimiser::Function, 
    model::LikelihoodModel, 
    edge_anti::Vector{Int}, 
    num_vertices::Int, 
    ind1::Int, 
    ind2::Int,
    biv_opt_is_ellipse_analytical::Bool, 
    ve1::Int, 
    ve2::Int, 
    relative_magnitude::Float64, 
    bound_warning::Bool, 
    save_internal_points::Bool)

Missing docstring.

Missing docstring for PlaceholderLikelihood.heapupdates_success!. Check Documenter's build log for details.

PlaceholderLikelihood.heapupdates_failure! — Function

This function is used in the event that no boundary points are found using newboundarypoint. Failure means it is likely that multiple level sets exist. If so, we break the edges of the candidate point and e_intersect and reconnect the vertexes such that we now have multiple boundary polygons.

If we only have one or two points on one of these boundary polygons we will display an info message as no additional points can be found from the method directly.

If we have three or more points on these boundary polygons, then there should be no problems finding other parts of these polygons.

If the largest polygon has less than two points the method will display a warning message and terminate, returning the boundary found up until then.

Missing docstring.

Missing docstring for PlaceholderLikelihood.polygon_break_and_rejoin!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.bivariate_confidenceprofile_iterativeboundary. Check Documenter's build log for details.

Vectorsearch Methods

For the RadialRandomMethod, RadialMLEMethod and SimultaneousMethod.

Missing docstring.

Missing docstring for PlaceholderLikelihood.generatepoint. Check Documenter's build log for details.

PlaceholderLikelihood.findNpointpairs_simultaneous! — Function

At later stage, implement with maxIters in the event that either the user-specified bounds don't contain internal points (or similarly external points) for a given confidence region (i.e. bounds either don't contain the 2D boundary at a confidence level OR the 2D boundary at a confidence level contains the bounds)

Missing docstring.

Missing docstring for PlaceholderLikelihood.find_m_spaced_radialdirections. Check Documenter's build log for details.

PlaceholderLikelihood.findNpointpairs_radialrandom! — Function

Distorts uniformly spaced angles on a circle to angles on an ellipse representative of the relative magnitude of each parameter. If the magnitude of a parameter is a NaN value (i.e. either bound is Inf), then the relative magnitude is set to 1.0, as no information is known about its magnitude.

Missing docstring.

Missing docstring for PlaceholderLikelihood.bivariate_confidenceprofile_vectorsearch. Check Documenter's build log for details.

Fix1Axis Method

For Fix1AxisMethod.

Missing docstring.

Missing docstring for PlaceholderLikelihood.findNpointpairs_fix1axis!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.bivariate_confidenceprofile_fix1axis!. Check Documenter's build log for details.

Continuation Method

For ContinuationMethod.

Missing docstring.

Missing docstring for PlaceholderLikelihood.update_targetll!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.normal_vector_i_2d!. Check Documenter's build log for details.

PlaceholderLikelihood.continuation_line_search! — Function

startlevelset only has two rows. Within this function we keep track of both the 2d points on the boundary - as this is all that needs to be known to get to the next point

For a given level set to get to, that's larger than all points in current level set.

Find normal direction at each point (either using ForwardDiff or a first order approximation)

Check where this normal direction intersects bounds and ensure that the next level set is bracketed by current point and point on boundary

If a bracket between current point and point on boundary does not exist the level set point recorded will be the point on the boundary

init guess for nuisance parameters should be their value at current point

using vector search bivariate function as input to find_zero(), using Order0 method, with starting guess of Ψ=0

record point at that level set

p and pointison_bounds is mutated by this function.

LATER: Update to allow specification of alternate method to find_zero(), e.g. Order8 OR alternate method using NLsolve (and LineSearches) in "2DGradientLinesearchTests.jl" which is gradient based.

Missing docstring.

Missing docstring for PlaceholderLikelihood.continuation_inwards_radial_search!. Check Documenter's build log for details.

PlaceholderLikelihood.initial_continuation_solution! — Function

The initial ellipse solution also should be in feasible region: contained within bounds specified for interest parameters. WARN if it is not - may cause some unexpected behaviour if the parameter is meant to be ≥ 0, yet is allowed to start there in the initial ellipse solution.

Find extrema of true log likelihoods of initial ellipse solution 3 cases: Case 1 is preferred. Warnings will be raised for both Case 2 and 3 Case 1: If min ll > than target ll of smallest target confidence level (preferred approach atm). Then line search from initial ellipse to ll boundary defined by min ll and this is the starting continuation solution. Line search in direction specified by forward diff at point

Case 2: If max ll < than target ll of smallest target confidence level. Line search radially towards the mle solution from the ellipse to the smallest target confidence level boundary and this is the starting continuation solution. In case two, this counts as a 'level set'.

Case 3: If min ll and max ll bracket the smallest target confidence level. Then line search radially towards the mle solution from initial ellipse to ll boundary defined by max ll and this is the starting continuation solution.

For both Case 2 and 3, could use gradient at initial elipse point rather than going radially towards the mle, but in my opinion may not be able to find the log likelihood boundary.

PlaceholderLikelihood.bivariate_confidenceprofile_continuation — Function

bivariateoptimiser is the optimiser to use with findzero bivariateoptimisergradient is the optimiser to use with ForwardDiff.gradient and to evaluate the obj value for the initial ellipse solution.

PlaceholderLikelihood.star_obj — Function

star_obj(centers, points)

centers have 2D points stored in columns.

points have 2D boundary points stored in columns and are in either clockwise or anticlockwise order of polygon connection.

PlaceholderLikelihood.boundary_smoother! — Function

smooth_boundary!(points, point_is_on_bounds)

points has 2D boundary points stored in columns and are in either clockwise or anticlockwise order of polygon connection. points is edited in place.

Points that we know are on the provided bounds are not moved/smoothed, although their presence will impact what the smoother tries to do.

Smoother must attempt to ensure that smoothed points don't go outside the level set boundary???? - may need to evaluate ll for every smoothed point to guarantee this.

PlaceholderLikelihood.refine_search_directions! — Function

Use kmeans with say 9 clusters on either: the polygon boundary points or points sampled within the polygon boundary.

Define objectives for the cluster points: how close each point is to being a star point, e.g. by finding which cluster point can see the most boundary points (i.e. there isn’t a edge in the way) - this is a form of regularisation by discretisation; we’re assuming that the boundary defined by a polygon with straight lines between boundary points is relatively consistent with the true shape. Additionally, we assume that our boundary points are relatively well spaced out so that the objective isn’t biased by a ton of our known points being located in a specific portion of the boundary.

If we find a star point (or multiple) then we should use that point to push out radially (e.g. instead of the MLE point). Note: If boundary is convex, all points in our set are star points by definition. If concave if star point exists use that for continuation, else use kmeans points which are likely to be star points for their local sections

Second obj we can use to tie break is how close each cluster point is to the centre of the polygon boundary (e.g. just using a simple how many points of all points (either the boundary points OR points that are sampled in the boundary) are on either side in x and y axis of the cluster point (and how close these are to 50%).

If star point doesn't exist then... Presently will use each individual kmeans point as the point to push out from based on the cluster the boundary point belongs to. In future perhaps use point closest to being a star point for all vertices it is a star point for?? And then other points for the others???

Returns whether or not a star point was found (if not, cannot guarantee that the ordering of boundary points will stay the same, and require a TSP iteration after solving for the next level set).