Common Functions

PlaceholderLikelihood.convertθnames_toindices

PlaceholderLikelihood.desired_df_subset

PlaceholderLikelihood.get_consistent_tuple

PlaceholderLikelihood.get_target_loglikelihood

PlaceholderLikelihood.ll_correction

PlaceholderLikelihood.convertθnames_toindices — Function

convertθnames_toindices(model::LikelihoodModel, 
    θnames_to_convert::Vector{<:Symbol})

Converts a vector of symbols representing parameters in model to a vector of each symbol's corresponding index in model.core.θnames.

convertθnames_toindices(model::LikelihoodModel, 
    θnames_to_convert::Union{Vector{Vector{Symbol}}, Vector{Tuple{Symbol, Symbol}}})

Converts a vector of vectors or tuples containing symbols representing parameters in model to a vector of vectors containing each symbol's corresponding index in model.core.θnames.

PlaceholderLikelihood.ll_correction — Function

ll_correction(model::LikelihoodModel, 
    profile_type::AbstractProfileType, 
    ll::Float64)

If a profile_type is LogLikelihood(), it corrects ll such that an input log-likelihood value (which has value of zero at the MLE) will now have a value of model.core.maximisedmle at the MLE. Otherwise, a copy of ll is returned, as both ellipse approximation profile types have a log-likelihood value of 0.0 at the MLE.

PlaceholderLikelihood.get_target_loglikelihood — Function

get_target_loglikelihood(model::LikelihoodModel, 
    confidence_level::Float64, 
    profile_type::AbstractProfileType, 
    dof::Int)

Returns the target log-likelihood / threshold at a confidence level and degrees of freedom, dof (the number of interest parameters), required for a particular profile_type to be in the confidence set. Uses PlaceholderLikelihood.ll_correction

PlaceholderLikelihood.get_consistent_tuple — Function

get_consistent_tuple(model::LikelihoodModel, 
    confidence_level::Float64, 
    profile_type::AbstractProfileType, 
    dof::Int)

Returns a tuple containing the values needed for log-likelihood evaluation and finding function zeros, including the target log-likelihood, number of model parameters, log-likelihood function to use and data tuple for evaluating the log-likelihood function.

PlaceholderLikelihood.desired_df_subset — Function

desired_df_subset(df::DataFrame, 
    num_used_rows::Int, 
    confidence_levels::Union{Float64, Vector{<:Float64}}, 
    sample_types::Vector{<:AbstractSampleType}; 
    sample_dimension::Int=0, 
    for_prediction_generation::Bool=false, 
    for_prediction_plots::Bool=false, 
    include_higher_confidence_levels::Bool=false)

Returns a view of df that includes only valid rows ∈ 1:num_used_rows, and rows that contain all of the values specified within function arguments. For dimensional samples.

Arguments

df: a DataFrame - model.dim_samples_df.

num_used_rows: the number of valid rows in df - model.num_dim_samples.

confidence_levels: a vector of confidence levels or a Float64 of a single confidence level. If empty, all confidence levels in df are allowed. Otherwise, if include_higher_confidence_levels == true and confidence_levels is a Float64, all confidence levels greater than or equal to confidence_levels are allowed. Else, only matching confidence levels in df are allowed.

sample_types: a vector of AbstractSampleType structs. If empty, all sample types in df are allowed. Otherwise, only matching sample types in df are allowed.

Keyword Arguments

sample_dimension: an integer greater than or equal to 0; if non-zero only matching dimensions of interest parameters in df are allowed, otherwise all are a allowed. Default is 0.

for_prediction_generation: a boolean specifying whether only rows which have not had predictions evaluated are allowed. As predictions do not need to be generated for rows which already have them evaluated.

for_prediction_plots: a boolean specifying whether only rows which have had predictions evaluated are allowed. As prediction plots can only include rows which have evaluated predictions.

include_higher_confidence_levels: a boolean specifying whether all confidence levels greater than or equal to confidence_levels are allowed. Useful for prediction plots as a dimensional sample can be evaluated at a high confidence level (e.g. 0.95) and then used at a lower confidence level (e.g. 0.9), extracting only the sample points that are in the 0.9 confidence set.

desired_df_subset(df::DataFrame, 
    num_used_rows::Int, 
    θs_of_interest::Vector{<:Int}, 
    confidence_levels::Union{Float64, Vector{<:Float64}}, 
    profile_types::Vector{<:AbstractProfileType}; 
    for_points_in_interval::Tuple{Bool,Int,Real}=(false,0,0), 
    for_prediction_generation::Bool=false, 
    for_prediction_plots::Bool=false)

Returns a view of df that includes only valid rows ∈ 1:num_used_rows, and rows that contain all of the values specified within function arguments. For univariate profiles.

Arguments

df: a DataFrame - model.uni_profiles_df.

num_used_rows: the number of valid rows in df - model.num_uni_profiles.

confidence_levels: a vector of confidence levels or a Float64 of a single confidence level. If empty, all confidence levels in df are allowed. Otherwise, if include_higher_confidence_levels == true and confidence_levels is a Float64, all confidence levels greater than or equal to confidence_levels are allowed. Else, only matching confidence levels in df are allowed.

profile_types: a vector of AbstractProfileType structs. If empty, all profile types in df are allowed. Otherwise, only matching profile types in df are allowed.

Keyword Arguments

for_points_in_interval: a tuple used for only extracting the rows that need to have points in the confidence interval evaluated by get_points_in_interval!. Default is (false, 0, 0).

for_prediction_generation: a boolean specifying whether only rows which have not had predictions evaluated are allowed. As predictions do not need to be generated for rows which already have them evaluated.

for_prediction_plots: a boolean specifying whether only rows which have had predictions evaluated are allowed. As prediction plots can only include rows which have evaluated predictions.

desired_df_subset(df::DataFrame, 
    num_used_rows::Int, 
    θs_of_interest::Vector{Tuple{Int,Int}}, 
    confidence_levels::Union{Float64, Vector{<:Float64}}, 
    profile_types::Vector{<:AbstractProfileType}, 
    methods::Vector{<:AbstractBivariateMethod}=AbstractBivariateMethod[]; 
    for_prediction_generation::Bool=false, 
    for_prediction_plots::Bool=false, 
    include_lower_confidence_levels::Bool=false)

Returns a view of df that includes only valid rows ∈ 1:num_used_rows, and rows that contain all of the values specified within function arguments. For bivariate profiles.

Arguments

df: a DataFrame - model.biv_profiles_df.

num_used_rows: the number of valid rows in df - model.num_biv_profiles.

confidence_levels: a vector of confidence levels or a Float64 of a single confidence level. If empty, all confidence levels in df are allowed. Otherwise, if include_higher_confidence_levels == true and confidence_levels is a Float64, all confidence levels greater than or equal to confidence_levels are allowed. Else, only matching confidence levels in df are allowed.

profile_types: a vector of AbstractProfileType structs. If empty, all profile types in df are allowed. Otherwise, only matching profile types in df are allowed.

methods: a vector of AbstractBivariateMethod structs. If empty, all methods in df are allowed. Otherwise, only methods of the same type in df are allowed.

Keyword Arguments

for_prediction_generation: a boolean specifying whether only rows which have not had predictions evaluated are allowed. As predictions do not need to be generated for rows which already have them evaluated.

for_prediction_plots: a boolean specifying whether only rows which have had predictions evaluated are allowed. As prediction plots can only include rows which have evaluated predictions.

include_lower_confidence_levels: a boolean specifying whether all confidence levels less than or equal to confidence_levels are allowed. Useful for prediction plots if a given set of bivariate profiles has few internal points evaluated, meaning some information about predictions may be missing. Bivariate profiles at lower confidence levels are by definition inside the desired confidence interval and may provide additional information on predictions.