Coverage Functions

LikelihoodBasedProfileWiseAnalysis.evaluate_conf_simultaneous_coverage

LikelihoodBasedProfileWiseAnalysis.evaluate_coverage

LikelihoodBasedProfileWiseAnalysis.evaluate_coverage_realisations

LikelihoodBasedProfileWiseAnalysis.evaluate_coverage_reference_sets

LikelihoodBasedProfileWiseAnalysis.union_of_prediction_extrema

LikelihoodBasedProfileWiseAnalysis.union_of_prediction_realisations_extrema

Predictions

LikelihoodBasedProfileWiseAnalysis.union_of_prediction_extrema — Function

union_of_prediction_extrema(df::DataFrame, dict::Dict)

Compute the extrema union of the extrema of the predictions of each profile in the rows of df, with extrema stored in dict.

LikelihoodBasedProfileWiseAnalysis.union_of_prediction_realisations_extrema — Function

union_of_prediction_realisations_extrema(df::DataFrame, dict::Dict)

Compute the extrema union of the extrema of the prediction realisations of each profile in the rows of df, with extrema stored in dict.

LikelihoodBasedProfileWiseAnalysis.evaluate_coverage — Function

evaluate_coverage(y_true::Array, 
    y_pred_extrema::Union{Array,Missing}, 
    multiple_outputs::Bool)

simultaneous coverage requires ytrue to be inside ypred for every row and column in y_true (every observed variable)

pointwise coverage is on a row and column basis; whether ytrue is inside ypred at each individual row and column

evaluate_coverage(model::LikelihoodModel, 
    y_true::Array, 
    profile_kind::Symbol, 
    multiple_outputs::Bool, 
    len_θs::Int)

LikelihoodBasedProfileWiseAnalysis.evaluate_coverage_realisations — Function

evaluate_coverage_realisations(model::LikelihoodModel, 
    testing_data::Array, 
    profile_kind::Symbol, 
    multiple_outputs::Bool, 
    len_θs::Int)

LikelihoodBasedProfileWiseAnalysis.evaluate_coverage_reference_sets — Function

evaluate_coverage_reference_sets(y_true::Array, 
    y_pred_extrema::Union{Array,Missing}, 
    multiple_outputs::Bool)

simultaneous coverage requires ytrue to be inside ypred for every row and column in y_true (every observed variable)

pointwise coverage is on a row and column basis; whether ytrue is inside ypred at each individual row and column

evaluate_coverage_reference_sets(model::LikelihoodModel, 
    reference_set_data::Tuple{Array,Array}, 
    profile_kind::Symbol, 
    multiple_outputs::Bool, 
    len_θs::Int, 
    not_missing_profiles::Bool)

LikelihoodBasedProfileWiseAnalysis.evaluate_conf_simultaneous_coverage — Function

evaluate_conf_simultaneous_coverage(pointwise::BitArray, 
    region::Float64)

Evaluate whether at least region proportion of entries in pointwise are true. This is simultaneous coverage with a different goal; i.e. region proportion of points are contained simultaneously.