Initialisation Internal

LikelihoodBasedProfileWiseAnalysis.calculate_θmagnitudes

LikelihoodBasedProfileWiseAnalysis.checkforInf

LikelihoodBasedProfileWiseAnalysis.init_biv_profile_row_exists!

LikelihoodBasedProfileWiseAnalysis.init_biv_profiles_df

LikelihoodBasedProfileWiseAnalysis.init_dim_samples_df

LikelihoodBasedProfileWiseAnalysis.init_dim_samples_row_exists!

LikelihoodBasedProfileWiseAnalysis.init_uni_profile_row_exists!

LikelihoodBasedProfileWiseAnalysis.init_uni_profiles_df

LikelihoodBasedProfileWiseAnalysis.init_uni_profiles_df — Function

init_uni_profiles_df(num_rows::Int; existing_largest_row::Int=0)

Initialises the DataFrame of model.uni_profiles_df with num_rows initial rows. In the event that the DataFrame already exists and more rows are being added, keyword argument, existing_largest_row, will be the number of rows in the existing dataframe, so that values of row_ind when concatenating the DataFrames will increase in steps of 1.

LikelihoodBasedProfileWiseAnalysis.init_biv_profiles_df — Function

init_biv_profiles_df(num_rows::Int; existing_largest_row::Int=0)

Initialises the DataFrame of model.biv_profiles_df with num_rows initial rows. In the event that the DataFrame already exists and more rows are being added, keyword argument, existing_largest_row, will be the number of rows in the existing dataframe, so that values of row_ind when concatenating the DataFrames will increase in steps of 1.

LikelihoodBasedProfileWiseAnalysis.init_dim_samples_df — Function

init_dim_samples_df(num_rows::Int; existing_largest_row::Int=0)

Initialises the DataFrame of model.dim_samples_df with num_rows initial rows. In the event that the DataFrame already exists and more rows are being added, keyword argument, existing_largest_row, will be the number of rows in the existing dataframe, so that values of row_ind when concatenating the DataFrames will increase in steps of 1.

LikelihoodBasedProfileWiseAnalysis.init_uni_profile_row_exists! — Function

init_uni_profile_row_exists!(model::LikelihoodModel, 
    θs_to_profile::Vector{<:Int}, 
    dof::Int,
    profile_type::AbstractProfileType)

Initialises the dictionary entry in model.uni_profile_row_exists for the key (θi, dof, profile_type), where θi is an element of θs_to_profile and dof is the degrees of freedom used to define the asymptotic threshold, with a DefaultDict with key of type Float64 (a confidence level) and default value of 0.

LikelihoodBasedProfileWiseAnalysis.init_biv_profile_row_exists! — Function

init_biv_profile_row_exists!(model::LikelihoodModel, 
    θcombinations::Vector{Vector{Int}}, 
    dof::Int,
    profile_type::AbstractProfileType, 
    method::AbstractBivariateMethod)

Initialises the dictionary entry in model.biv_profile_row_exists for the key ((ind1, ind2), dof, profile_type, method), where (ind1, ind2) is a combination in θcombinations and dof is the degrees of freedom used to define the asymptotic threshold, with a DefaultDict with key of type Float64 (a confidence level) and default value of 0.

LikelihoodBasedProfileWiseAnalysis.init_dim_samples_row_exists! — Function

init_dim_samples_row_exists!(model::LikelihoodModel, 
    sample_type::AbstractSampleType)

Initialises the dictionary entry in model.dim_samples_row_exists for the key (sample_type) with a DefaultDict with key of type Float64 (a confidence level) and default value of 0. For a full likelihood sample (dimension equal to the number of model parameters).

init_dim_samples_row_exists!(model::LikelihoodModel, 
    θindices::Vector{Vector{Int}},
    sample_type::AbstractSampleType)

Initialises the dictionary entry in model.dim_samples_row_exists for the key (θvec, dof, sample_type), where θvec is a vector in θindices and dof=length(θvec) is the degrees of freedom used to define the asymptotic threshold, with a DefaultDict with key of type Float64 (a confidence level) and default value of 0. For a non-full likelihood sample (dimension less than the number of model parameters).

Estimating Parameter Magnitudes and Bounds

LikelihoodBasedProfileWiseAnalysis.calculate_θmagnitudes — Function

calculate_θmagnitudes(θlb::Vector{<:Float64}, θub::Vector{<:Float64})

Estimates the magnitude for each parameter using the difference between parameter bounds. If a bound is an Inf, the value is set to NaN. Values are divided by the minimum estimated magnitude such that the returned magnitudes have a lowest value of 1.0.

Parameter Transformations

LikelihoodBasedProfileWiseAnalysis.checkforInf — Function

checkforInf(x::AbstractVector{<:Real})

Warns via a message if any of the bounds returned given the provided forward transformation are +/-Inf.

Arguments

x: vector of transformed bounds.