Plotting Functions

LikelihoodBasedProfileWiseAnalysis.addMLE!

LikelihoodBasedProfileWiseAnalysis.addMLEandLLstar!

LikelihoodBasedProfileWiseAnalysis.add_extrema!

LikelihoodBasedProfileWiseAnalysis.add_yMLE!

LikelihoodBasedProfileWiseAnalysis.plot1Dprofile!

LikelihoodBasedProfileWiseAnalysis.plot2Dboundary!

LikelihoodBasedProfileWiseAnalysis.plotprediction!

LikelihoodBasedProfileWiseAnalysis.plotrealisation!

LikelihoodBasedProfileWiseAnalysis.profile1Dlinestyle

LikelihoodBasedProfileWiseAnalysis.profile2Dmarkershape

LikelihoodBasedProfileWiseAnalysis.profilecolor

LikelihoodBasedProfileWiseAnalysis.θcombinations_to_plot_typeconversion

LikelihoodBasedProfileWiseAnalysis.θs_to_plot_typeconversion

Helper Functions

LikelihoodBasedProfileWiseAnalysis.profilecolor — Function

profilecolor(profile_type::Union{AbstractProfileType, AbstractSampleType})

Colors of profile types and sample types as integers between 1 and 6.

LikelihoodBasedProfileWiseAnalysis.profile1Dlinestyle — Function

profile1Dlinestyle(profile_type::AbstractProfileType)

Linestyle of each profile_type. EllipseApproxAnalytical - :dash, EllipseApprox - :dashdot, LogLikelihood - :solid,

LikelihoodBasedProfileWiseAnalysis.profile2Dmarkershape — Function

profile2Dmarkershape(profile_type::Union{AbstractProfileType, AbstractSampleType}, on_boundary::Bool)

Marker shapes of bivariate profiles depending on profile type or sample type and whether or not the point is on the profile's boundary.

LikelihoodBasedProfileWiseAnalysis.θs_to_plot_typeconversion — Function

θs_to_plot_typeconversion(model::LikelihoodModel, θs_to_plot::Union{Vector{<:Symbol}, Vector{<:Int64}})

Converts θs_to_plot to integer index values with LikelihoodBasedProfileWiseAnalysis.convertθnames_toindices if they are supplied as a vector of symbols.

LikelihoodBasedProfileWiseAnalysis.θcombinations_to_plot_typeconversion — Function

θcombinations_to_plot_typeconversion(model::LikelihoodModel,
    θcombinations_to_plot::Union{Vector{Vector{Symbol}}, Vector{Tuple{Symbol, Symbol}}, Vector{Vector{Int}}, Vector{Tuple{Int,Int}}})

Converts θcombinations_to_plot to a vector of tuples of integer index values with LikelihoodBasedProfileWiseAnalysis.convertθnames_toindices if they are supplied as a vector of vectors/tuples containing symbols. The index elements in each vector/tuple will be sorted in ascending order.

Plots

LikelihoodBasedProfileWiseAnalysis.plot1Dprofile! — Function

plot1Dprofile!(plt, parRange, parProfile, label="profile"; kwargs...)

Plots a univariate profile at ψ locations parRange, with normalised profile log-likelihood function values parProfile.

LikelihoodBasedProfileWiseAnalysis.addMLEandLLstar! — Function

addMLEandLLstar!(plt, llstar, parMLE, MLE_color, llstar_color; kwargs...)

On a univariate profile, adds the MLE location as a vertical line and the asymptotic confidence threshold as a horizontal line. The intersection between the horizontal asymptotic confidence threshold and the univariate profile gives the location of the corresponding confidence interval for that parameter.

LikelihoodBasedProfileWiseAnalysis.plot2Dboundary! — Function

plot2Dboundary!(plt, parBoundarySamples, label="boundary"; use_lines=false, kwargs...)

Plots the boundary of a bivariate profile at 2D locations parBoundarySamples. If use_lines=false then it will use a scatter plot, otherwise it will connect the boundary in the order of the parBoundarySamples.

LikelihoodBasedProfileWiseAnalysis.addMLE! — Function

addMLE!(plt, parMLEs; kwargs...)

On a bivariate profile, adds the MLE location as a scatter point.

LikelihoodBasedProfileWiseAnalysis.plotprediction! — Function

plotprediction!(plt, t, predictions, extrema, linealpha, layout; extremacolor=:red, kwargs...)

Plots the profile-wise confidence set for the model trajectory using the same format as in the profile-wise analysis workflow paper [1]. The extrema of the profile-wise trajectory confidence set is labelled as approximate profile-wise simultaneous confidence bands (SCBs).

LikelihoodBasedProfileWiseAnalysis.plotrealisation! — Function

plotrealisation!(plt, t, extrema, linealpha, layout; extremacolor=:red, kwargs...)

Plots the extrema of the profile-wise reference tolerance set for the (1-δ) population reference set. The (1-δ) population reference set refers to a set containing the (1-δ) reference region of the population at each time point; i.e. the smallest region at each time point which contains (1-δ) of possible realisations. The extrema of the profile-wise reference tolerance set is labelled as approximate profile-wise simultaneous reference tolerance bands (SRTBs).

LikelihoodBasedProfileWiseAnalysis.add_yMLE! — Function

add_yMLE!(plt, t, yMLE, layout; kwargs...)

Adds the model trajectory obtained from simulating the model at times t using the maximum likelihood estimate for parameters.

LikelihoodBasedProfileWiseAnalysis.add_extrema! — Function

add_extrema!(plt, t, extrema, layout; extremacolor=:gold, label=["Sampled SCBs (≈)" ""])

Adds additional extrema to plots for the model trajectory and population reference sets. Typically for comparing the extrema of profile-wise predictions sets from profiles with sets from dimensional samples (often the full parameter confidence set).