Prediction Functions

LikelihoodBasedProfileWiseAnalysis.generate_prediction

LikelihoodBasedProfileWiseAnalysis.generate_prediction_bivariate

LikelihoodBasedProfileWiseAnalysis.generate_prediction_univariate

LikelihoodBasedProfileWiseAnalysis.predict_realisations

LikelihoodBasedProfileWiseAnalysis.generate_prediction — Function

generate_prediction(predictfunction::Function,
    errorfunction::Function,
    data,
    t::AbstractVector,
    data_ymle::AbstractArray{<:Real},
    parameter_points::Matrix{Float64},
    proportion_to_keep::Real,
    region::Real,
    channel::Union{RemoteChannel,Missing}=missing)

Generates the predictions for response variables from a predictfunction which meets the requirements specified in add_prediction_function!, given data, at time points t for each parameter combination in the columns of parameter_points. The extrema of all predictions is computed and proportion_to_keep of the individual predictions are kept. errorfunction is used to predict the lower and upper quartiles of realisations at each prediction point; the highest density region is returned. In particular it's predicting the region populuation reference interval at each prediction point. When computed from a parameter confidence set like a profile, the extrema gives the extrema of the profile-wise (region, confidence_level) reference tolerance set.

Returns a [PredictionStruct] containing the kept predictions, prediction extrema, lower and upper quartiles of realisations from the error model of region at each predicted point and the realisation extrema.

The prediction at each timepoint is stored in the corresponding row (1st dimension). The prediction for each parameter combination is stored in the corresponding column (2nd dimension). The prediction for multiple response variables is stored in the 3rd dimension.

generate_prediction(predictfunction::Function,
    errorfunction::Missing,
    data,
    t::AbstractVector,
    data_ymle::AbstractArray{<:Real},
    parameter_points::Matrix{Float64},
    proportion_to_keep::Real,
    region::Real,
    channel::Union{RemoteChannel,Missing}=missing)

Generates the predictions for response variables from a predictfunction which meets the requirements specified in add_prediction_function!, given data, at time points t for each parameter combination in the columns of parameter_points. The extrema of all predictions is computed and proportion_to_keep of the individual predictions are kept.

Returns a [PredictionStruct] containing the kept predictions and prediction extrema.

The prediction at each timepoint is stored in the corresponding row (1st dimension). The prediction for each parameter combination is stored in the corresponding column (2nd dimension). The prediction for multiple response variables is stored in the 3rd dimension.

LikelihoodBasedProfileWiseAnalysis.generate_prediction_univariate — Function

generate_prediction_univariate(model::LikelihoodModel,
    errorfunction::Union{Function, Missing},
    sub_df,
    row_i::Int,
    t::AbstractVector,
    proportion_to_keep::Real,  
    channel::RemoteChannel)

Generates predictions for the univariate profile in sub_df that corresponds to row_i at timepoints t.

LikelihoodBasedProfileWiseAnalysis.generate_prediction_bivariate — Function

generate_prediction_bivariate(model::LikelihoodModel,
    sub_df,
    row_i::Int,
    t::AbstractVector,
    proportion_to_keep::Real, 
    channel::RemoteChannel)

Generates predictions for the bivariate profile in sub_df that corresponds to row_i at timepoints t.

LikelihoodBasedProfileWiseAnalysis.predict_realisations — Function

predict_realisations(errorfunction::Function, 
    predictions::AbstractArray, 
    θ::AbstractVector,
    region::Float64)

Uses errorfunction to make prediction for realisations; it forms region reference intervals for region population reference intervals. Returns two arrays of the same dimension as predictions, one containing the lower quantile and one containing the upper quantile (the extrema) of the reference interval.