PlaceholderLikelihood.get_points_in_interval_single_row

PlaceholderLikelihood.univariate_confidenceinterval

Univariate Functions

Initialisation and Array Mapping

Missing docstring.

Missing docstring for PlaceholderLikelihood.variablemapping1dranges. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.variablemapping1d!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.boundsmapping1d!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.init_univariate_parameters. Check Documenter's build log for details.

Likelihood Optimisation

Missing docstring.

Missing docstring for PlaceholderLikelihood.univariateΨ_ellipse_unbounded. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.univariateΨ. Check Documenter's build log for details.

Get Points in Confidence Interval

Missing docstring.

Missing docstring for PlaceholderLikelihood.update_uni_dict_internal!. Check Documenter's build log for details.

PlaceholderLikelihood.get_points_in_interval_single_row — Function

Will get points in the interval and at the interval boundaries - currentintervalpoints has at least the boundary points.

additional_width is the additional width past the interval boundary to also evaluate points on.

Main Confidence Interval Logic

Missing docstring.

Missing docstring for PlaceholderLikelihood.get_interval_brackets. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.add_uni_profiles_rows!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.set_uni_profiles_row!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.get_univariate_opt_func. Check Documenter's build log for details.

PlaceholderLikelihood.univariate_confidenceinterval — Function

mle_targetll means that the variable is created such that the ll at the mle is 0.0

Missing docstring.

Missing docstring for PlaceholderLikelihood.univariate_confidenceinterval_master. Check Documenter's build log for details.