Predictions

PlaceholderLikelihood.AbstractPredictionStruct

PlaceholderLikelihood.PredictionStruct

PlaceholderLikelihood.add_prediction_function!

PlaceholderLikelihood.check_prediction_function_exists

PlaceholderLikelihood.generate_predictions_univariate!

In the event that a prediction function has not been added to the LikelihoodModel struct yet, we can add one using add_prediction_function!

PlaceholderLikelihood.add_prediction_function! — Function

add_prediction_function!(model::LikelihoodModel, predictfunction::Function)

Adds a prediction function, predictfunction, to model.

Requirements for predictfunction: a function to generate model predictions from that is paired with the loglikefunction. Takes three arguments, θ, data and t, in that order, where θ and data are the same as for loglikefunction and t needs to be an optional third argument. When t is not specified, the prediction function should be evaluated for the same time points/independent variable as the data. When t is specified, the prediction function should be evaluated for those specified time points/independent variable.

PlaceholderLikelihood.check_prediction_function_exists — Function

check_prediction_function_exists(model::LikelihoodModel)

Checks if a prediction function is stored in model.

Then to generate predictions we can use one of three functions, depending on whether we want to generate predictions from univariate or bivariate profiles, or dimensional samples.

PlaceholderLikelihood.generate_predictions_univariate! — Function

Missing docstring.

Missing docstring for generate_predictions_bivariate!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for generate_predictions_dim_samples!. Check Documenter's build log for details.

Structs

PlaceholderLikelihood.AbstractPredictionStruct — Type

AbstractPredictionStruct

Supertype for the predictions storage struct.

Subtypes

PredictionStruct

PlaceholderLikelihood.PredictionStruct — Type

PredictionStruct(predictions::Array{Real}, extrema::Array{Real})

Struct for containing evaluated predictions corresponding to confidence profiles.

Fields

predictions: array of model predictions evaluated at the parameters given by a particular confidence profile parameter set. If a model has multiple predictive variables, it assumes that model.predictfunction stores the prediction for each variable in its columns. We are going to store values for each variable in the 3rd dimension (row=dim1, col=dim2, page/sheet=dim3). Each column corresponds to a column of the confidence profile parameter set.

extrema: extrema of the predictions array.

Supertype Hiearachy

PredictionStruct <: AbstractPredictionStruct <: Any