Ellipse Functions

LikelihoodBasedProfileWiseAnalysis.analytic_ellipse_loglike

LikelihoodBasedProfileWiseAnalysis.analytic_ellipse_loglike_1D_soln

LikelihoodBasedProfileWiseAnalysis.ellipse_like

LikelihoodBasedProfileWiseAnalysis.ellipse_loglike

LikelihoodBasedProfileWiseAnalysis.getMLE_hessian_and_covariance

LikelihoodBasedProfileWiseAnalysis.test_hessian_identifiability

LikelihoodBasedProfileWiseAnalysis.getMLE_hessian_and_covariance — Function

getMLE_hessian_and_covariance(f::Function, θmle::Vector{<:Float64})

Computes the negative hessian of function f at θmle using ForwardDiff.jl and it's pseudoinverse, returning both matrices.

Hessian identifiability is tested using LikelihoodBasedProfileWiseAnalysis.test_hessian_identifiability.

LikelihoodBasedProfileWiseAnalysis.test_hessian_identifiability — Function

test_hessian_identifiability(Hmle::Matrix{T}, num_pars::Int) where T<:Float64

Modified R code from Cole, (2020): https://doi.org/10.1201/9781315120003, (https://www.kent.ac.uk/smsas/personal/djc24/parameterredundancy.html) in book code, bassicocc.R.

The cutoff used to estimate whether a standardised eigenvalue is close enough to zero to indicate non-identifiability is 1e-12*number of parameters, which is a smaller value than used in Viallefont et. al. (1998) (https://doi.org/10.1002/(SICI)1521-4036(199807)40:3<313::AID-BIMJ313>3.0.CO;2-2) and Cole (2020) (https://doi.org/10.1201/9781315120003) due to smaller error in the calculation of Hmle via automatic differentiation. This cutoff is meant as an indication of non-identifiability/singularity; the hessian may still be identifiable/non-singular even if a warning occurs.

Likelihood

LikelihoodBasedProfileWiseAnalysis.analytic_ellipse_loglike — Function

Analytic ellipse log-likelihood has no knowledge of lower and upper bounds on parameters. Hence profiles generated by optimising out the nuisance parameters for a given interest parameter may look different if the analytical profile enters space where a bound would be active. Pushing forward from these confidence bounds may be infeasible - if analytical profile has entered a space where a parameter bound is active.

LikelihoodBasedProfileWiseAnalysis.analytic_ellipse_loglike_1D_soln — Function

analytic_ellipse_loglike_1D_soln(θIndex::Int, mleTuple::@NamedTuple{θmle::Vector{T}, Γmle::Matrix{T}}, targetll::T)

See Structural and practical identifiability analysis of partially observed dynamical models by exploiting the profile likelihood, equation 7.

\[L^* = -\frac{1}{2}(θ_i - θ_i^*)^2 \times Γ_{ii}(θ^*)^{-1}\]

\[θ_i = θ_i^* + \sqrt{\frac{-2 L^*}{Γ_{ii}(θ^*)^{-1}}} \equiv θ_i^* + \sqrt{-2 L^* \times Γ_{ii}(θ^*)}\]

Note: $C(θ^*) = 2 \times H(θ^*)^{-1}$ and $Γ(θ^*) = H(θ^*)^{-1}$, and $L^* = -χ^2(α, df)$, so the equation is equivalent to equation 7 in the above reference.

LikelihoodBasedProfileWiseAnalysis.ellipse_loglike — Function

ellipse_loglike(θ::Vector, mleTuple::@NamedTuple{θmle::Vector{T}, Hmle::Matrix{T}}) where T<:Float64

LikelihoodBasedProfileWiseAnalysis.ellipse_like — Function

ellipse_like(θ::Vector{T}, mleTuple::@NamedTuple{θmle::Vector{T}, Hmle::Matrix{T}}) where T<:Float64