Ellipse Functions

PlaceholderLikelihood.analytic_ellipse_loglike

PlaceholderLikelihood.analytic_ellipse_loglike_1D_soln

Missing docstring.

Missing docstring for PlaceholderLikelihood.getMLE_hessian_and_covariance. Check Documenter's build log for details.

Likelihood

PlaceholderLikelihood.analytic_ellipse_loglike — Function

Analytic ellipse log-likelihood has no knowledge of lower and upper bounds on parameters. Hence profiles generated by optimising out the nuisance parameters for a given interest parameter may look different if the analytical profile enters space where a bound would be active. Pushing forward from these confidence bounds may be infeasible - if analytical profile has entered a space where a parameter bound is active.

PlaceholderLikelihood.analytic_ellipse_loglike_1D_soln — Function

\[L^* = -\frac{1}{2}(w_i - w_i^*)^2 * Γ_i(θ^*)^{-1} w_1 = w_1^* + \sqrt{\frac{-2 L^*}{Γ_i(θ^*)^{-1}}}\]

Missing docstring.

Missing docstring for PlaceholderLikelihood.ellipse_loglike. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.ellipse_like. Check Documenter's build log for details.