Univariate Functions

PlaceholderLikelihood.add_uni_profiles_rows!

PlaceholderLikelihood.get_interval_brackets

PlaceholderLikelihood.get_points_in_interval_single_row

PlaceholderLikelihood.get_univariate_opt_func

PlaceholderLikelihood.set_uni_profiles_row!

PlaceholderLikelihood.univariate_confidenceinterval

PlaceholderLikelihood.update_uni_dict_internal!

Initialisation and Array Mapping

Missing docstring.

Missing docstring for PlaceholderLikelihood.variablemapping1dranges. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.variablemapping1d!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.boundsmapping1d!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.init_univariate_parameters. Check Documenter's build log for details.

Likelihood Optimisation

Missing docstring.

Missing docstring for PlaceholderLikelihood.univariateΨ_ellipse_unbounded. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PlaceholderLikelihood.univariateΨ. Check Documenter's build log for details.

Get Points in Confidence Interval

PlaceholderLikelihood.update_uni_dict_internal! — Function

update_uni_dict_internal!(model::LikelihoodModel,
    uni_row_number::Int,
    points::PointsAndLogLikelihood)

Updates the interval_points field of a UnivariateConfidenceStruct, for the profile related to uni_row_number stored at model.uni_profiles_dict[uni_row_number], with the interval points stored in points.

PlaceholderLikelihood.get_points_in_interval_single_row — Function

get_points_in_interval_single_row(univariate_optimiser::Function, 
    model::LikelihoodModel,
    num_points_in_interval::Int,
    θi::Int,
    profile_type::AbstractProfileType,
    current_interval_points::PointsAndLogLikelihood,
    additional_width::Real=0.0)

Method for getting num_points_in_interval points inside a confidence interval for parameter θi, directly called by PlaceholderLikelihood.univariate_confidenceinterval and called via it's other method for get_points_in_interval. Optionally adds additional_width outside of the confidence interval, so long as a parameter bound is not reached. If a bound is reached, up until the bound will be considered instead.

get_points_in_interval_single_row(model::LikelihoodModel,
    uni_row_number::Int,
    num_points_in_interval::Int,
    additional_width::Real)

Alternate method called by get_points_in_interval.

Main Confidence Interval Logic

PlaceholderLikelihood.get_interval_brackets — Function

get_interval_brackets(model::LikelihoodModel, 
    θi::Int, 
    confidence_level::Float64, 
    profile_type::AbstractProfileType)

Returns updated interval brackets (Float64 vectors of length two) if smaller or larger confidence level profiles exist for θi, such that the region to bracket over for the left and right sides of the confidence internval is smallest. Otherwise, returns empty brackets (Float64[]).

PlaceholderLikelihood.add_uni_profiles_rows! — Function

add_uni_profiles_rows!(model::LikelihoodModel, num_rows_to_add::Int)

Adds num_rows_to_add free rows to model.uni_profiles_df by vertically concatenating the existing DataFrame and free rows using PlaceholderLikelihood.init_uni_profiles_df.

PlaceholderLikelihood.set_uni_profiles_row! — Function

set_uni_profiles_row!(model::LikelihoodModel, 
    row_ind::Int, 
    θi::Int,
    not_evaluated_internal_points::Bool, 
    not_evaluated_predictions::Bool,
    confidence_level::Float64, 
    profile_type::AbstractProfileType, 
    num_points::Int, 
    additional_width::Real)

Sets the relevant fields of row row_ind in model.uni_profiles_df after a profile has been evaluated.

PlaceholderLikelihood.get_univariate_opt_func — Function

get_univariate_opt_func(profile_type::AbstractProfileType=LogLikelihood())

Returns the correct univariate optimisation function used to for find the optimal values of nuisance parameters for a given interest parameter value for the profile_type log-likelihood function. The optimisation function returns the value of the profile_type log-likelihood function as well as finding the optimal nuisance parameters and saving these in one of it's inputs.

Will be PlaceholderLikelihood.univariateΨ for the LogLikelihood() and EllipseApprox() profiles types and PlaceholderLikelihood.univariateΨ_ellipse_unbounded for the EllipseApproxAnalytical profile type.

PlaceholderLikelihood.univariate_confidenceinterval — Function

univariate_confidenceinterval!(p::Progress,
    univariate_optimiser::Function, 
    model::LikelihoodModel, 
    consistent::NamedTuple, 
    θi::Int, 
    confidence_level::Float64,
    profile_type::AbstractProfileType,
    mle_targetll::Float64,
    use_existing_profiles::Bool,
    num_points_in_interval::Int,
    additional_width::Real)

Returns a UnivariateConfidenceStruct containing the likelihood-based confidence interval for interest parameter θi at confidence_level, and any additional points within the interval if num_points_in_interval > 0 as well as outside the interval if num_points_in_interval > 0 and additional_width > 0. Log-likelihood values, standardised to 0.0 at the MLE point, for all points found in the interval are also stored in the UnivariateConfidenceStruct.

If use_existing_profiles=true then the brackets used to find each side of the confidence interval (between each side of the bounds for θi and the MLE point), will be updated and made smaller if confidence profiles for θi already exist at lower and higher confidence levels.

Called by univariate_confidenceintervals!.

Missing docstring.

Missing docstring for PlaceholderLikelihood.univariate_confidenceinterval_master. Check Documenter's build log for details.