Initialisation

PlaceholderLikelihood.CoreLikelihoodModel

PlaceholderLikelihood.EllipseMLEApprox

PlaceholderLikelihood.LikelihoodModel

PlaceholderLikelihood.initialiseLikelihoodModel

PlaceholderLikelihood.optimise

PlaceholderLikelihood.optimise_unbounded

PlaceholderLikelihood.setbounds!

PlaceholderLikelihood.setmagnitudes!

PlaceholderLikelihood.transformbounds

PlaceholderLikelihood.transformbounds_NLopt

PlaceholderLikelihood.initialiseLikelihoodModel — Function

initialiseLikelihoodModel(loglikefunction::Function,
    predictfunction::Union{Function, Missing},
    data::Union{Tuple, NamedTuple},
    θnames::Vector{<:Symbol},
    θinitialGuess::AbstractVector{<:Real},
    θlb::AbstractVector{<:Real},
    θub::AbstractVector{<:Real},
    θmagnitudes::AbstractVector{<:Real}=Float64[];
    uni_row_prealloaction_size=NaN,
    biv_row_preallocation_size=NaN,
    dim_row_preallocation_size=NaN,
    show_progress=true)

Initialises a LikelihoodModel struct, which contains all model information, profiles, samples and predictions.

Arguments

loglikefunction: a loglikelihood function which takes two arguments, θ and data, in that order, where θ is a vector containing the values of each parameter in θnames and data is a Tuple or NamedTuple - see data below.

predictfunction: a prediction function to generate model predictions from that is paired with the loglikefunction. Takes three arguments, θ, data and t, in that order, where θ and data are the same as for loglikefunction and t needs to be an optional third argument. When t is not specified, the prediction function should be evaluated for the same time points/independent variable as the data. When t is specified, the prediction function should be evaluated for those specified time points/independent variable. It can also be missing if no function is provided to initialiseLikelihoodModel, because predictions are not required when evaluating parameter profiles. The function can be added at a later point using add_prediction_function!.

data: a Tuple or a NamedTuple containing any additional information required by the loglikelihood function, such as the time points to be evaluated at.

θnames: a vector of symbols containing the names of each parameter, e.g. [:λ, :K, :C0].

θinitialguess: a vector containing the initial guess for the values of each parameter. Used to find the MLE point.

θlb: a vector of lower bounds on parameters.

θub: a vector of upper bounds on parameters.

θmagnitudes: a vector of the relative magnitude of each parameter. If not provided, it will be estimated using the difference of θlb and θub with PlaceholderLikelihood.calculate_θmagnitudes. Can be updated after initialisation using setθmagnitudes!.

Keyword Arguments

uni_row_prealloaction_size: number of rows of uni_profiles_df to preallocate. Default is NaN (a single row).

biv_row_preallocation_size:number of rows of biv_profiles_df to preallocate. Default is NaN (a single row).

dim_row_preallocation_size: number of rows of dim_samples_df to preallocate. Default is NaN (a single row).

show_progress: Whether to show the progress of profiling across sets of interest parameters.

initialiseLikelihoodModel(loglikefunction::Function,
    data::Union{Tuple, NamedTuple},
    θnames::Vector{<:Symbol},
    θinitialGuess::Vector{<:Float64},
    θlb::Vector{<:Float64},
    θub::Vector{<:Float64},
    θmagnitudes::Vector{<:Real}=zeros(0);
    uni_row_prealloaction_size=NaN,
    biv_row_preallocation_size=NaN,
    dim_row_preallocation_size=NaN,
    show_progress=true)

Alternate version of initialiseLikelihoodModel that can be called without a prediction function. The function can be added at a later point using add_prediction_function!.

PlaceholderLikelihood.LikelihoodModel — Type

LikelihoodModel(core::CoreLikelihoodModel, 
    ellipse_MLE_approx::Union{Missing, EllipseMLEApprox},
    num_uni_profiles::Int, 
    num_biv_profiles::Int, 
    num_dim_samples::Int, 
    uni_profiles_df::DataFrame, 
    biv_profiles_df::DataFrame, 
    dim_samples_df::DataFrame, 
    uni_profile_row_exists::Dict{Tuple{Int, AbstractProfileType}, DefaultDict{Float64, Int}},
    biv_profile_row_exists::Dict{Tuple{Tuple{Int, Int}, AbstractProfileType, AbstractBivariateMethod}, DefaultDict{Float64, Int}},
    dim_samples_row_exists::Dict{Union{AbstractSampleType, Tuple{Vector{Int}, AbstractSampleType}}, DefaultDict{Float64, Int}},
    uni_profiles_dict::Dict{Int, UnivariateConfidenceStruct}, 
    biv_profiles_dict::Dict{Int, BivariateConfidenceStruct}, 
    dim_samples_dict::Dict{Int, SampledConfidenceStruct},
    uni_predictions_dict::Dict{Int, AbstractPredictionStruct}, 
    biv_predictions_dict::Dict{Int, AbstractPredictionStruct}, 
    dim_predictions_dict::Dict{Int, AbstractPredictionStruct},
    show_progress::Bool)

Struct containing all the information required to compute profiles, samples and predictions. Created by initialiseLikelihoodModel.

Fields

core: a [CoreLikelihoodModel] struct.

ellipse_MLE_approx: a [EllipseMLEApprox] struct OR a missing value if the ellipse approximation of the log-likelihood at the MLE point has not been evaluated yet.

num_uni_profiles: the number of different univariate profiles that have been evaluated (distinct combinations of different confidence levels, AbstractProfileType structs and single interest parameters). Specifies the number of valid rows in uni_profiles_df.

num_biv_profiles: the number of different bivariate profiles that have been evaluated (distinct combinations of different confidence levels, AbstractProfileType structs, AbstractBivariateMethod structs and two interest parameters). Specifies the number of valid rows in biv_profiles_df.

num_dim_samples: the number of different dimensional profiles that have been evaluated (distinct combinations of different confidence levels, AbstractProfileType structs, AbstractSampleType structs and sets of interest parameters). Specifies the number of valid rows in dim_samples_df.

uni_profiles_df: a DataFrame with each row containing information on each univariate profile evaluated, where the row index is the key for that profile in uni_profiles_dict and uni_predictions_dict.

biv_profiles_df: a DataFrame with each row containing information on each bivariate profile evaluated, where the row index is the key for that profile in biv_profiles_dict and biv_predictions_dict.

dim_samples_df: a DataFrame with each row containing information on each dimensional sample evaluated, where the row index is the key for that sample in dim_samples_dict and dim_predictions_dict.

uni_profile_row_exists: a dictionary containing information on whether a row in uni_profiles_df exists for a given combination of interest parameter, AbstractProfileType and confidence level. If it does exist, it's value will be the row index in uni_profiles_df otherwise it will be 0.

biv_profile_row_exists: a dictionary containing information on whether a row in biv_profiles_df exists for a given combination of two interest parameters, AbstractProfileType, AbstractBivariateMethod and confidence level. If it does exist, it's value will be the row index in biv_profiles_df otherwise it will be 0.

dim_samples_row_exists: a dictionary containing information on whether a row in dim_samples_df exists for a given combination of interest parameter, AbstractProfileType, AbstractSampleType and confidence level. If it does exist, it's value will be the row index in dim_samples_df otherwise it will be 0.

uni_profiles_dict: a dictionary with keys of type Integer and values of type [UnivariateConfidenceStruct] containing the profile for each valid row in uni_profiles_df. The row index of uni_profiles_df is the key for the corresponding profile.

biv_profiles_dict: a dictionary with keys of type Integer and values of type [BivariateConfidenceStruct] containing the profile for each valid row in biv_profiles_df. The row index of biv_profiles_df is the key for the corresponding profile.

dim_samples_dict: a dictionary with keys of type Integer and values of type [SampledConfidenceStruct] containing the profile for each valid row in dim_samples_df. The row index of dim_samples_df is the key for the corresponding profile.

uni_predictions_dict: a dictionary with keys of type Integer and values of type PredictionStruct containing the predictions from the profiles in uni_profiles_dict for each valid row in uni_profiles_df. The row index of uni_profiles_df is the key for the corresponding prediction, if that prediction has been calculated using generate_predictions_univariate!.

biv_predictions_dict: a dictionary with keys of type Integer and values of type PredictionStruct containing the predictions from the profiles in biv_profiles_dict for each valid row in biv_profiles_df. The row index of biv_profiles_df is the key for the corresponding prediction, if that prediction has been calculated using generate_predictions_bivariate!.

dim_predictions_dict: a dictionary with keys of type Integer and values of type PredictionStruct containing the predictions from the profiles in dim_samples_dict for each valid row in dim_samples_df. The row index of dim_samples_df is the key for the corresponding prediction, if that prediction has been calculated using generate_predictions_dim_samples!.

show_progress: a boolean specifying whether to show the progress of profile methods with respect to sets of interest parameter(s).

Supertype Hiearachy

LikelihoodModel <: Any

PlaceholderLikelihood.CoreLikelihoodModel — Type

CoreLikelihoodModel(loglikefunction::Function, 
    predictfunction::Union{Function, Missing}, 
    data::Union{Tuple, NamedTuple}, 
    θnames::Vector{<:Symbol}, 
    θname_to_index::Dict{Symbol, Int}, 
    θlb::AbstractVector{<:Real}, 
    θub::AbstractVector{<:Real}, 
    θmagnitudes::AbstractVector{<:Real}, 
    θmle::Vector{<:Float64}, 
    ymle::Array{<:Real}, 
    maximisedmle::Float64, 
    num_pars::Int)

Struct containing the core information required to define a LikelihoodModel. For additional information on parameters (where repeated), see initialiseLikelihoodModel.

Fields

loglikefunction: a log-likelihood function which takes two arguments, θ and data, in that order.

predictfunction: a prediction function to generate model predictions from that is paired with the loglikefunction.

data: a Tuple or a NamedTuple containing any additional information required by the log-likelihood function, such as the time points to be evaluated at.

θnames: a vector of symbols containing the names of each parameter, e.g. [:λ, :K, :C0].

θname_to_index: a dictionary with keys of type Symbol and values of type Int, with the key being an element of θnames and the value being the corresponding index of the key in θnames.

θlb: a vector of lower bounds on parameters.

θub: a vector of upper bounds on parameters.

θmagnitudes: a vector of the relative magnitude of each parameter.

θmle: a vector containing the maximum likelihood estimate for each parameter.

ymle: an array containing the output of the prediction function at θmle and data.

maximisedmle: the value of the log-likelihood function evaluated at θmle.

num_pars: the number of model parameters, length(θnames).

Supertype Hiearachy

CoreLikelihoodModel <: Any

Ellipse Approximation

Missing docstring.

Missing docstring for getMLE_ellipse_approximation!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for check_ellipse_approx_exists!. Check Documenter's build log for details.

PlaceholderLikelihood.EllipseMLEApprox — Type

EllipseMLEApprox(Hmle::Matrix{<:Float64}, Γmle::Matrix{<:Float64})

Struct containing two n*n arrays representing the ellipse approximation of the log-likelihood function around the MLE point. See getMLE_ellipse_approximation!

Fields

Hmle: a n*n array, where n is the number of model parameters, containing the negative Hessian of the log-likelihood function evaluated at the MLE point.

Γmle: a n*n array, where n is the number of model parameters, containing the inverse of Hmle.

Supertype Hiearachy

EllipseMLEApprox <: Any

Modifying Parameter Magnitudes and Bounds

PlaceholderLikelihood.setmagnitudes! — Function

setmagnitudes!(model::LikelihoodModel, θmagnitudes::AbstractVector{<:Real})

Updates the magnitudes of each parameter in model from model.core.θmagnitudes to θmagnitudes.

PlaceholderLikelihood.setbounds! — Function

setbounds!(model::LikelihoodModel; 
    lb::AbstractVector{<:Real}=Float64[], 
    ub::AbstractVector{<:Real}=Float64[])

Updates the parameter bounds in model from model.core.θlb to lb if specified and from model.core.θub to ub if specified. lb and ub are keyword arguments.

Parameter Transformations

PlaceholderLikelihood.transformbounds — Function

transformbounds(transformfun::Function, 
    lb::AbstractVector{<:Real}, 
    ub::AbstractVector{<:Real}, 
    independentParameterIndexes::Vector{<:Int}=Int[], 
    dependentParameterIndexes::Vector{<:Int}=Int[])

parameters are vectors of ints - i.e. call using vectors of ints directly or look up position of parameter from a symbol vector using a lookup table. Note. we assume that ordering remains the same. A 'independentParameter' is one where the new parameter Θ[i] depends only on f(θ[i]). A 'dependentParameter' is one where the new parameter Θ[i] depends on f(θ[i], θ[j], j!=i).

I suspect that the dependentParameter heuristic may fail if there are multiple local minima - a binary integer programme may be required instead (however, integer requirement on variables can be relaxed)

ONLY VALID FOR MONOTONIC (increasing or decreasing) TRANSFORMATIONS OF VARIABLES

PlaceholderLikelihood.transformbounds_NLopt — Function

transformbounds_NLopt(transformfun::Function, 
    lb::AbstractVector{<:Real}, 
    ub::AbstractVector{<:Real})

Optimisation of Nuisance Parameters

This is presently done using a predefined NLopt optimiser.

PlaceholderLikelihood.optimise — Function

optimise(fun, θ₀, lb, ub;
    dv = false, method = dv ? :LD_LBFGS : :LN_BOBYQA)

NLopt optimiser used for calculating the values of nuisance parameters.

PlaceholderLikelihood.optimise_unbounded — Function

optimise_unbounded(fun, θ₀;
    dv = false, method = dv ? :LD_LBFGS : :LN_BOBYQA)

Alternative version of optimise without nuisance parameter bounds. Used for computing the nuisance parameters of EllipseApproxAnalytical profiles.