Bivariate Profiles

PlaceholderLikelihood.AbstractBivariateMethod

PlaceholderLikelihood.AbstractBivariateVectorMethod

PlaceholderLikelihood.AnalyticalEllipseMethod

PlaceholderLikelihood.ContinuationMethod

PlaceholderLikelihood.Fix1AxisMethod

PlaceholderLikelihood.IterativeBoundaryMethod

PlaceholderLikelihood.RadialMLEMethod

PlaceholderLikelihood.RadialRandomMethod

PlaceholderLikelihood.SimultaneousMethod

PlaceholderLikelihood.bivariate_confidenceprofiles!

PlaceholderLikelihood.bivariate_methods

PlaceholderLikelihood.sample_bivariate_internal_points!

PlaceholderLikelihood.bivariate_confidenceprofiles! — Function

bivariate_confidenceprofiles!(model::LikelihoodModel, 
    θcombinations::Vector{Vector{Int}}, 
    num_points::Int; 
    <keyword arguments>)

Finds num_points profile_type boundary points at a specified confidence_level for each combination of two interest parameters using a specified method, optionally saving any found internal points. Saves these profiles by modifying model in place.

Arguments

model: a LikelihoodModel containing model information, saved profiles and predictions.

θcombinations: vector of pairs of parameters to profile, as a vector of vectors of model parameter indexes.

num_points: number of points to find on the boundary at the specified confidence level. Depending on the method, if a region of the user-provided bounds is inside the boundary some of these points will be on the bounds and inside the boundary.

Keyword Arguments

confidence_level: a number ∈ (0.0, 1.0) for the confidence level on which to find the profile_type boundary. Default is 0.95 (95%).

profile_type: whether to use the true log-likelihood function or an ellipse approximation of the log-likelihood function centred at the MLE (with optional use of parameter bounds). Available profile types are LogLikelihood, EllipseApprox and EllipseApproxAnalytical. Default is LogLikelihood() (LogLikelihood).

method: a method of type AbstractBivariateMethod. For a list of available methods use bivariate_methods() (bivariate_methods). Default is RadialRandomMethod(3) (RadialRandomMethod).

θcombinations_is_unique: boolean variable specifying whether all parameter combinations in θcombinations are ordered by parameter index (ascending) and are unique. Default is false.

save_internal_points: boolean variable specifying whether to save points found inside the boundary during boundary computation. Internal points can be plotted in bivariate profile plots and will be used to generate predictions from a given bivariate profile. Default is true.

existing_profiles: Symbol ∈ [:ignore, :merge, :overwrite] specifying what to do if profiles already exist for a given θcombination, confidence_level, profile_type and method. See below for each symbol's meanings. Default is :merge.

show_progress: boolean variable specifying whether to display progress bars on the percentage of θcombinations completed and estimated time of completion. Default is model.show_progress.

use_distributed: boolean variable specifying whether to use a normal for loop or a @distributed for loop across combinations of interest parameters. The variable makes no difference if Distributed.jl is not being used - it's intended for use when simulating parameter confidence boundary coverage (see check_bivariate_parameter_coverage).

existing_profiles meanings

:ignore means profiles that already exist will not be recomputed even if they contain fewer num_points boundary points.

:merge means profiles that already exist will be merged with profiles from the current algorithm run to reach num_points. If the existing profile already has at least num_points boundary points then that profile will not be recomputed. Otherwise, the specified method will be run starting from the difference between num_points and the number of points in the existing profile. The result of that method run will be merged with the existing profile.

:overwrite means profiles that already exist will be overwritten, regardless of how many points they contain.

Details

Using bivariate_confidenceprofile this function calls the algorithm/method specified by method for each parameter combination in θcombinations (depending on the setting for existing_profiles and num_points if these profiles already exist). Updates model.biv_profiles_df for each successful profile and saves their results as a BivariateConfidenceStruct in model.biv_profiles_dict, where the keys for the dictionary is the row number in model.biv_profiles_df of the corresponding profile.

bivariate_confidenceprofiles!(model::LikelihoodModel, 
    θcombinations_symbols::Union{Vector{Vector{Symbol}}, Vector{Tuple{Symbol, Symbol}}}, 
    num_points::Int; 
    <keyword arguments>)

Profiles just the provided θcombinations_symbols parameter pairs, provided as either a vector of vectors or a vector of tuples.

bivariate_confidenceprofiles!(model::LikelihoodModel, 
    profile_m_random_combinations::Int, 
    num_points::Int; 
    <keyword arguments>)

Profiles m random two-way combinations of model parameters (sampling without replacement), where 0 < m ≤ binomial(model.core.num_pars,2).

θcombinations_is_unique keyword argument

θcombinations_is_unique is not a valid keyword argument for this function method as it internally produces the parameter combinations which are guaranteed to be unique.

bivariate_confidenceprofiles!(model::LikelihoodModel, 
    num_points::Int; 
    <keyword arguments>)

Profiles all two-way combinations of model parameters.

θcombinations_is_unique keyword argument

θcombinations_is_unique is not a valid keyword argument for this function method as it internally produces the parameter combinations which are guaranteed to be unique.

Sampling Internal Points

PlaceholderLikelihood.sample_bivariate_internal_points! — Function

sample_bivariate_internal_points!(model::LikelihoodModel,
    num_points::Int;
    confidence_levels::Vector{<:Float64}=Float64[],
    profile_types::Vector{<:AbstractProfileType}=AbstractProfileType[],
    methods::Vector{<:AbstractBivariateMethod}=AbstractBivariateMethod[],
    t::Union{Vector, Missing}=missing,
    evaluate_predictions_for_samples::Bool=true,
    proportion_of_predictions_to_keep::Real=1.0)

Methods

PlaceholderLikelihood.AbstractBivariateMethod — Type

AbstractBivariateMethod

Supertype for bivariate boundary finding methods. Use bivariate_methods() for a list of available methods (see bivariate_methods).

Subtypes

AbstractBivariateVectorMethod

Fix1AxisMethod

ContinuationMethod

AnalyticalEllipseMethod

PlaceholderLikelihood.AbstractBivariateVectorMethod — Type

AbstractBivariateVectorMethod <: AbstractBivariateMethod

Supertype for bivariate boundary finding methods that search between two distinct points.

Subtypes

IterativeBoundaryMethod

RadialMLEMethod

RadialRandomMethod

SimultaneousMethod

Supertype Hiearachy

AbstractBivariateVectorMethod <: AbstractBivariateMethod <: Any

PlaceholderLikelihood.bivariate_methods — Function

bivariate_methods()

Prints a list of available bivariate methods. Available bivariate methods include IterativeBoundaryMethod, RadialRandomMethod, RadialMLEMethod, SimultaneousMethod, Fix1AxisMethod, ContinuationMethod and AnalyticalEllipseMethod.

PlaceholderLikelihood.IterativeBoundaryMethod — Type

IterativeBoundaryMethod(initial_num_points::Int, 
    angle_points_per_iter::Int, 
    edge_points_per_iter::Int, 
    radial_start_point_shift::Float64=rand(), 
    ellipse_sqrt_distortion::Float64=1.0;
    use_ellipse::Bool=false)

Method for iteratively improving an initial boundary of initial_num_points, found by pushing out from the MLE point in directions defined by either RadialMLEMethod, use_ellipse=true, or [RadialRandomMethod], use_ellipse=false (see PlaceholderLikelihood.findNpointpairs_radialMLE!, PlaceholderLikelihood.iterativeboundary_init and PlaceholderLikelihood.bivariate_confidenceprofile_iterativeboundary).

Arguments

initial_num_points: a positive integer number of initial boundary points to find.

angle_points_per_iter: a integer ≥ 0 for the number of edges to explore and improve based on the angle objective before switching to the edge objective. If angle_points_per_iter > 0 and edge_points_per_iter > 0 the angle objective is considered first.

edge_points_per_iter: a integer ≥ 0 for the number of edges to explore and improve based on the edge objective before switching back to the angle objective. If angle_points_per_iter > 0 and edge_points_per_iter > 0 the angle objective is considered first.

radial_start_point_shift: a number ∈ [0.0,1.0]. Default is rand() (defined on [0.0,1.0]), meaning that by default a different set of initial points will be found each time.

ellipse_sqrt_distortion: a number ∈ [0.0,1.0]. Default is 0.01.

Keyword Arguments

use_ellipse: Whether to find initial_num_points by searching in directions defined by an ellipse approximation, as in RadialMLEMethod, or , as in RadialRandomMethod. Default is false.

Details

For additional information on the radial_start_point_shift and ellipse_sqrt_distortion arguments see the keyword arguments for generate_N_clustered_points in EllipseSampling.jl.

Recommended for use with the LogLikelihood profile type. Radial directions, edge length and internal angle calculations are rescaled by the relative magnitude/scale of the two interest parameters. This is so directions and regions explored and consequently the boundary found are not dominated by the parameter with the larger magnitude.

Once an initial boundary is found by pushing out from the MLE point in directions defined by either RadialMLEMethod or [RadialRandomMethod], the method seeks to improve this boundary by minimising an internal angle and an edge length objective, each considered sequentially, until the desired number of boundary points are found. As such, the method can be thought of as a mesh smoothing or improvement algorithm; we can consider the boundary found at a given moment in time to be a N-sided polygon with edges between adjacent boundary points (vertices).

Regions we define as needing improvement are those with:

Internal angles between adjacent edges that are far from being straight (180 degrees or π radians). The region defined by these edges is not well explored as the real boundary edge in this region likely has some degree of smooth curvature. It may also indicate that one of these edges cuts our boundary into a enclosed region and an unexplored region on the other side of the edge. In the event that a vertex is on a user-provided bound for a parameter, this objective is set to zero, as this angle region is a byproduct of user input and not the actual log-likelihood region. This objective is defined in PlaceholderLikelihood.internal_angle_from_pi!.

Edges between adjacent vertices that have a large euclidean distance. The regions between these vertices is not well explored as it is unknown whether the boundary actually is straight between these vertices. On average the closer two vertices are, the more likely the edge between the two points is well approximated by a straight line, and thus our mesh is a good representation of the true log-likelihood boundary. This objective is defined in PlaceholderLikelihood.edge_length.

The method is implemented as follows:

Create edges between adjacent vertices on the intial boundary. Calculate angle and edge length objectives for these edges and vertices and place them into tracked heaps.

Until found the desired number of boundary points repeat steps 3 and 4.

For angle_points_per_iter points:

Peek at the top vertex of the angle heap.

Place a candidate point in the middle of the edge connected to this vertex that has the larger angle at the vertex the edge connects to. This is so we explore the worse defined region of the boundary.

Use this candidate to find a new boundary point (see below).

If found a new boundary point, break edge we placed candidate on and replace with edges to the new boundary point, updating angle and edge length objectives in the tracked heap (see PlaceholderLikelihood.heapupdates_success!). Else break our polygon into multiple polygons (see PlaceholderLikelihood.heapupdates_failure).

For edge_points_per_iter points:

Peek at the top edge of the edge heap.

Place a candidate point in the middle of this edge.

Same as for step 3.3.

Same as for step 3.4.

angle_points_per_iter and edge_points_per_iter

At least one of angle_points_per_iter and edge_points_per_iter must be non-zero.

Using a candidate point to find a new boundary point

Uses PlaceholderLikelihood.newboundarypoint!.

If a candidate point is on the log-likelihood threshold boundary, we accept the point.

If a candidate point is inside the boundary, then we search in the normal direction to the edge until we find a boundary point or hit the parameter bound, accepting either.

If a candidate point is outside the boundary we find the edge, e_intersect of our boundary polygon that is intersected by the line in the normal direction of the candidate edge, which passes through the candidate point. Once this edge is found, we find the vertex on e_intersect that is closest to our candidate point. We setup a 1D line search/bracketing method between these two points. In the event that no boundary points are found between these two points it is likely that multiple boundaries exist. If so, we break the candidate point's edge and e_intersect and reconnect the vertexes such that we now have multiple boundary polygons.

Largest boundary polygon at any iteration must have at least three points

If the largest polygon has less than two points the method will display a warning message and terminate, returning the boundary found up until then.

Boundary finding method

For the initial boundary, it uses a 1D bracketing algorithm between the MLE point and the point on the user-provided bounds in the search directions to find the boundary at the desired confidence level. For boundary improvement, if the candidate point is inside the boundary, it uses a 1D bracketing algorithm between an internal point and the point on the user-provided bounds in the search direction. If the candidate point is outside the boundary, it uses a derivative-free 1D line search algorithm in the search direction. If the derivative-free approach fails, it switches to a bracketing heuristic between the candidate point and closest vertex on e_intersect.

This method is unlikely to find boundaries that do not contain the MLE point (if they exist), but it can find them. If a boundary that does not contain the MLE point is found, it is not guaranteed to be explored. In this case the the method will inform the user that multiple boundaries likely exist for this combination of model parameters.

Impact of parameter bounds

If a parameter bound is in the way of reaching the boundary in a given search direction, the point on that bound will be returned as the boundary point. For the bracketing method to work, parameter values on the bounds need to be legal and return a non Inf value from the log-likelihood function. Interest parameter bounds that have ranges magnitudes larger than the range of the boundary may prevent the true boundary from being found. Additionally, the bracketing method will be less efficient if the interest parameter bounds are far from the true boundary.

Internal Points

Finds between 0 and num_points - initial_num_points internal points - internal points are found when the edge being considered's midpoint is inside the boundary.

Supertype Hiearachy

IterativeBoundaryMethod <: AbstractBivariateVectorMethod <: AbstractBivariateMethod <: Any

PlaceholderLikelihood.RadialMLEMethod — Type

RadialMLEMethod(ellipse_start_point_shift::Float64=rand(), 
    ellipse_sqrt_distortion::Float64=0.01)

Method for finding the bivariate boundary of a confidence profile by bracketing between the MLE point and points on the provided bounds in directions given by points found on the boundary of a ellipse approximation of the log-likelihood function around the MLE, e, using EllipseSampling.jl (see PlaceholderLikelihood.findNpointpairs_radialMLE! and PlaceholderLikelihood.bivariate_confidenceprofile_vectorsearch)..

Arguments

ellipse_start_point_shift: a number ∈ [0.0,1.0]. Default is rand() (defined on [0.0,1.0]), meaning that by default a different set of points will be found each time.

ellipse_sqrt_distortion: a number ∈ [0.0,1.0]. Default is 0.01.

Details

For additional information on arguments see the keyword arguments for generate_N_clustered_points in EllipseSampling.jl.

Recommended for use with the EllipseApprox profile type. Will produce reasonable results for the LogLikelihood profile type when bivariate profile boundaries are convex. Otherwise, IterativeBoundaryMethod, which has an option to use a starting solution from RadialMLEMethod, is preferred as it iteratively improves the quality of the boundary and can discover regions not explored by this method.

Boundary finding method

Uses a 1D bracketing algorithm between the MLE point and the point on the user-provided bounds in the search directions to find the boundary at the desired confidence level.

This method is unlikely to find boundaries that do not contain the MLE point (if they exist).

Impact of parameter bounds

If a parameter bound is in the way of reaching the boundary in a given search direction, the point on that bound will be returned as the boundary point. For the bracketing method to work, parameter values on the bounds need to be legal and return a non Inf value from the log-likelihood function. Interest parameter bounds that have ranges magnitudes larger than the range of the boundary may prevent the true boundary from being found. Additionally, the bracketing method will be less efficient if the interest parameter bounds are far from the true boundary.

Internal Points

Finds no internal points.

Supertype Hiearachy

RadialMLEMethod <: AbstractBivariateVectorMethod <: AbstractBivariateMethod <: Any

PlaceholderLikelihood.RadialRandomMethod — Type

RadialRandomMethod(num_radial_directions::Int)

Method for finding the bivariate boundary of a confidence profile by finding internal boundary points using a uniform random distribution on provided bounds and choosing num_radial_directions to explore from these points (see PlaceholderLikelihood.findNpointpairs_radialRandom! and PlaceholderLikelihood.bivariate_confidenceprofile_vectorsearch).

Arguments

num_radial_directions: an integer greater than zero.

Details

Recommended for use with the LogLikelihood profile type. Radial directions are rescaled by the relative magnitude/scale of the two interest parameters. This is so directions explored and consequently the boundary found are not dominated by the parameter with the larger magnitude.

The method first uniformly samples the region specified by the bounds for the two interest parameters until a point within the boundary is found. Then it chooses num_radial_directions, spaced uniformly around a circle, and rotates these directions by a random phase shift ∈ [0.0, 2π/num_directions] radians. These directions are then distorted by the relative magnitude/scale of the two interest parameters. Then for each of these directions, until it runs out of directions or finds the desired number of points, it finds the closest point on the bounds from the internal point in this direction. If the point on the bounds is outside the boundary, it accepts the point pair. A bracketing method is then used to find a boundary point between the point pair (the bound point and the internal point). The method continues searching for internal points and generating directions until the desired number of boundary points is found.

Given a fixed number of desired boundary points, we can decrease the cost of finding internal points by increasing the number of radial directions to explore, num_radial_directions, at each internal point. However, it is important to balance num_radial_directions with the desired number of boundary points. If num_radial_directions is large relative to the number of boundary points, then the boundary the method finds may constitute a local search around the found internal points. Resultantly, there may be regions were the boundary is not well explored. This will be less of an issue for more 'circular' boundary regions.

IterativeBoundaryMethod may be preferred over this method if evaluating the log-likelihood function is expensive or the bounds provided for the interest parameters are much larger than the boundary, as the uniform random sampling for internal points will become very expensive.

Boundary finding method

Uses a 1D bracketing algorithm between an internal point and the point on the user-provided bounds in the search direction to find the boundary at the desired confidence level.

This method can find multiple boundaries (if they exist).

Impact of parameter bounds

If a parameter bound is in the way of reaching the boundary in a given search direction, in contrast to RadialMLEMethod, the point on that bound will not be returned as the boundary point. For the bracketing method to work, parameter values on the bounds need to be legal and return a non Inf value from the log-likelihood function. Interest parameter bounds that have ranges magnitudes larger than the range of the boundary may prevent the true boundary from being found. Additionally, the bracketing method will be less efficient if the interest parameter bounds are far from the true boundary. The method will fail if the interest parameter bounds are fully contained by the boundary.

Internal Points

Finds a minimum of div(num_points, num_radial_directions, RoundUp) internal points.

Supertype Hiearachy

RadialRandomMethod <: AbstractBivariateVectorMethod <: AbstractBivariateMethod <: Any

PlaceholderLikelihood.SimultaneousMethod — Type

SimultaneousMethod()

Method for finding the bivariate boundary of a confidence profile by finding internal and external boundary points using a uniform random distribution on provided bounds, pairing these points in the order they're found and bracketing between each pair (see PlaceholderLikelihood.findNpointpairs_simultaneous! and PlaceholderLikelihood.bivariate_confidenceprofile_vectorsearch).

Details

Recommended for use with the LogLikelihood profile type.

The method uniformly samples the region specified by the bounds for the two interest parameters until as many internal and external boundary points as the desired number of boundary points are found. These points are paired in the order they are found. A bracketing method is then used to find a boundary point between the point pair (the external point and the internal point).

RadialRandomMethod and IterativeBoundaryMethod are preferred over this method from a computational performance standpoint as they require fewer log-likelihood evalutions (when RadialRandomMethod has parameter num_radial_directions > 1).

Boundary finding method

Uses a 1D bracketing algorithm between a valid point pair to find the boundary at the desired confidence level.

This method can find multiple boundaries (if they exist).

Impact of parameter bounds

If a parameter bound is in the way of reaching the boundary, points will not be put on that bound. Additionally, if the true boundary is very close to a parameter bound, the method will struggle to find this region of the boundary. This is because finding the boundary in this location requires generating a random point between the boundary and the parameter bound, which becomes more difficult the closer they are. Interest parameter bounds that have ranges magnitudes larger than the range of the boundary will make finding internal points very difficult, requiring a lot of computational effort. Similarly, the inverse will be true if external points are hard to find. The method will fail if the interest parameter bounds are fully contained by the boundary.

Internal Points

Finds num_points internal points.

Supertype Hiearachy

SimultaneousMethod <: AbstractBivariateVectorMethod <: AbstractBivariateMethod <: Any

PlaceholderLikelihood.Fix1AxisMethod — Type

Fix1AxisMethod()

Method for finding the bivariate boundary of a confidence profile by using uniform random distributions on provided bounds to draw a value for one interest parameter, fix it, and draw two values for the other interest parameter, using the pair to find a boundary point using a bracketing method if the pair are a valid bracket (see PlaceholderLikelihood.findNpointpairs_fix1axis! and PlaceholderLikelihood.bivariate_confidenceprofile_fix1axis).

Details

Recommended for use with the LogLikelihood profile type.

The method finds the desired number of boundary points by fixing the first and second interest parameters for half of these points each. It first draws a value for the fixed parameter using a uniform random distribution on provided bounds (e.g. ψx). Then it draws two values for the unfixed parameter in the same fashion (e.g. ψy1 and ψy2]). If one of these points ([ψx, ψy1] and [ψx, ψ_y2]) is an internal point and the other an external point, the point pair is accepted as they are a valid bracket. A bracketing method is then used to find a boundary point between the point pair (the internal and external point). The method continues searching for valid point pairs until the desired number of boundary points is found.

RadialRandomMethod and IterativeBoundaryMethod are preferred over this method from a computational performance standpoint as they require fewer log-likelihood evalutions (when RadialRandomMethod has parameter num_radial_directions > 1). SimultaneousMethod will also likely be more efficient, even though it uses four random numbers vs three, as it doesn't unneccesarily throw away points.

Boundary finding method

Uses a 1D bracketing algorithm between a valid point pair, which are parallel to the x or y axis, to find the boundary at the desired confidence level.

This method can find multiple boundaries (if they exist).

Impact of parameter bounds

If a parameter bound is in the way of reaching the boundary, points will not be put on that bound. Additionally, if the true boundary is very close to a parameter bound, the method will struggle to find this region of the boundary. This is because finding the boundary in this location requires generating a random point between the boundary and the parameter bound, which becomes more difficult the closer they are. Interest parameter bounds that have ranges magnitudes larger than the range of the boundary will make finding internal points very difficult, requiring a lot of computational effort. Similarly, the inverse will be true if external points are hard to find. The method will fail if the interest parameter bounds are fully contained by the boundary.

Internal Points

Finds num_points internal points.

Supertype Hiearachy

Fix1AxisMethod <: AbstractBivariateMethod <: Any

PlaceholderLikelihood.AnalyticalEllipseMethod — Type

AnalyticalEllipseMethod(ellipse_start_point_shift::Float64, 
    ellipse_sqrt_distortion::Float64)

Method for sampling the desired number of boundary points on a ellipse approximation of the log-likelihood function centred at the maximum likelihood estimate point using EllipseSampling.jl.

Arguments

ellipse_start_point_shift: a number ∈ [0.0,1.0]. Default is rand() (defined on [0.0,1.0]), meaning that by default a different set of points will be found each time.

ellipse_sqrt_distortion: a number ∈ [0.0,1.0]. Default is 1.0, meaning that by default points on the ellipse approximation are equally spaced with respect to arc length.

Details

Used for the EllipseApproxAnalytical profile type only: if this method is specified, then any user provided profile type will be overriden and replaced with EllipseApproxAnalytical. This ellipse approximation ignores user provided bounds.

For additional information on arguments see the keyword arguments for generate_N_clustered_points in EllipseSampling.jl.

Boundary finding method

Explicitly finds the boundary using EllipseSampling.jl.

Internal Points

Finds no internal points.

Supertype Hiearachy

AnalyticalEllipseMethod <: AbstractBivariateMethod <: Any

PlaceholderLikelihood.ContinuationMethod — Type

ContinuationMethod(num_level_sets::Int, 
    ellipse_confidence_level::Float64, 
    ellipse_start_point_shift::Float64=rand(), 
    level_set_spacing::Symbol=:loglikelihood)

Kept available for completeness but not recommended for use. A previous implementation of search directions from the MLE point was moved to [RadialMLEMethod].

Arguments

num_level_sets: the number of level sets used to get to the desired confidence level set.

ellipse_confidence_level: a number ∈ (0.0, 1.0). the confidence level at which to construct the initial ellipse and find the initial level set boundary. Recommended to be around 0.1.

ellipse_start_point_shift: a number ∈ [0.0,1.0]. Default is rand() (defined on [0.0,1.0]), meaning that by default a different set of points will be found each time.

ellipse_sqrt_distortion: a number ∈ [0.0,1.0]. Default is 1.0, meaning that by default points on the ellipse approximation are equally spaced with respect to arc length.

level_set_spacing: a Symbol ∈ [:loglikelihood, :confidence]. Whether to space level sets uniformly in confidence level space or log-likelihood space, between the first level set found and the level set of desired confidence level. Default is :loglikelihood.

Details

The method finds an initial boundary at a low confidence level close to the ellipse_confidence_level (see initial_continuation_solution!). Then it 'continues' this initial boundary sequentially to num_level_sets higher confidence level boundaries until the desired confidence level is reached.

Presently this continuation is done by finding a point inside the boundary that is as close as possible to being a point that makes the boundary a star domain and is close to the centre of the area of the boundary in the x and y axes. We refer to this point as a 'star point', or a point that can see all the points on the boundary, without being blocked by an edge. We use a heuristic to estimate this, sampling points within the boundary and using these to produce kmeans, of which one is hopefully a star point and at the centre of the boundary.

If we find a star point we then, for every point on the current boundary, push out in the direction defined by the line segment connecting the star point and the boundary to find the next confidence level boundary. If we do not find a star point, we assign each of the boundary points to the Kmeans point they are closest to (using a Euclidean distance metric), and use the direction defined by the line segments between a boundary point and it's associated Kmeans point to find the next confidence level boundary. This direction heuristic is carried out by refine_search_directions!.

A traveling salesman heuristic is used to reorder the boundary points into a new minimum perimeter polygon, PlaceholderLikelihood.minimum_perimeter_polygon!, if the continuation of one boundary to the next causes the mapping of adjacent vertices to change (expected if a star point is not found).

For additional information on the ellipse_start_point_shift and ellipse_sqrt_distortion arguments see the keyword arguments for generate_N_clustered_points in EllipseSampling.jl.

Boundary finding method

Uses a derivative-free 1D line search algorithm to find the boundary at subsequent confidence levels. If the derivative-free approach fails, it switches to a bracketing algorithm between a given boundary point and the point on the user-provided bounds in the search direction.

This method is unlikely to find boundaries that do not contain the MLE point (if they exist).

Internal Points

Finds num_points * num_level_sets internal points at distinct level sets.

Supertype Hiearachy

ContinuationMethod <: AbstractBivariateMethod <: Any