Dimensional Samples

PlaceholderLikelihood.AbstractSampleType

PlaceholderLikelihood.LatinHypercubeSamples

PlaceholderLikelihood.UniformGridSamples

PlaceholderLikelihood.UniformRandomSamples

PlaceholderLikelihood.dimensional_likelihood_sample!

PlaceholderLikelihood.full_likelihood_sample!

We can generate samples within the loglikelihood boundary at any dimension of model interest parameters.

Sample Types

PlaceholderLikelihood.AbstractSampleType — Type

AbstractSampleType

Supertype for sample types.

Subtypes

UniformGridSamples

UniformRandomSamples

LatinHypercubeSamples

PlaceholderLikelihood.UniformGridSamples — Type

UniformGridSamples()

Evaluate the parameter bounds space on a uniform grid and keep samples that are inside the particular profile_type's boundary.

Supertype Hiearachy

UniformGridSamples <: AbstractSampleType <: Any

PlaceholderLikelihood.UniformRandomSamples — Type

UniformRandomSamples()

Take uniform random samples of parameter bounds space and keep samples that are inside the particular profile_type's boundary.

Supertype Hiearachy

UniformRandomSamples <: AbstractSampleType <: Any

PlaceholderLikelihood.LatinHypercubeSamples — Type

LatinHypercubeSamples()

Create a Latin Hypercube sampling plan in parameter bounds space and keep samples that are inside the particular profile_type's boundary. Uses LatinHypercubeSampling.jl.

Supertype Hiearachy

LatinHypercubeSamples <: AbstractSampleType <: Any

Full Likelihood Sampling

PlaceholderLikelihood.full_likelihood_sample! — Function

full_likelihood_sample!(model::LikelihoodModel,
    num_points_to_sample::Union{Int, Vector{Int}};
    confidence_level::Float64=0.95,
    sample_type::AbstractSampleType=LatinHypercubeSamples(),
    lb::AbstractVector{<:Real}=Float64[],
    ub::AbstractVector{<:Real}=Float64[],
    use_threads::Bool=true,
    existing_profiles::Symbol=:overwrite)

Dimensional Likelihood Sampling

Note: dimensional likelihood samples can be 'full' likelihood samples as well.

PlaceholderLikelihood.dimensional_likelihood_sample! — Function

dimensional_likelihood_sample!(model::LikelihoodModel,
    θindices::Vector{Vector{Int}},
    num_points_to_sample::Union{Int, Vector{Int}};
    confidence_level::Float64=0.95,
    sample_type::AbstractSampleType=LatinHypercubeSamples(),
    lb::AbstractVector{<:Real}=Float64[],
    ub::AbstractVector{<:Real}=Float64[],
    θs_is_unique::Bool=false,
    use_threads::Bool=true,
    existing_profiles::Symbol=:overwrite,
    show_progress::Bool=model.show_progress)

Missing docstring.

Missing docstring for bivariate_concave_hull. Check Documenter's build log for details.