Structs and Profile Types

PlaceholderLikelihood.AbstractConfidenceStruct

PlaceholderLikelihood.AbstractEllipseProfileType

PlaceholderLikelihood.AbstractProfileType

PlaceholderLikelihood.BivariateConfidenceStruct

PlaceholderLikelihood.EllipseApprox

PlaceholderLikelihood.EllipseApproxAnalytical

PlaceholderLikelihood.LogLikelihood

PlaceholderLikelihood.PointsAndLogLikelihood

PlaceholderLikelihood.SampledConfidenceStruct

PlaceholderLikelihood.UnivariateConfidenceStruct

Structs

PlaceholderLikelihood.PointsAndLogLikelihood — Type

PointsAndLogLikelihood(points::Array{Float64}, 
    ll::Vector{<:Float64}, 
    boundary_col_indices::Vector{<:Int64}=zeros(Int, 0)))

Struct that stores an array of parameter points, their corresponding log-likelihood value and, in the case of univariate profiles, the column indices in points of the confidence interval parameters.

Fields

points: an array of points stored in columns, with each row corresponding to the respective index of each model parameter. For the UnivariateConfidenceStruct type, these points are stored in column-wise order of increasing interest parameter magnitude. For the BivariateConfidenceStruct type these points are stored in the order they are found.

ll: a vector of log-likelihood function values corresponding to the point in each column of points. This number is standardised so that regardless of whether the true log-likelihood function or an ellipse approximation of the function is evaluated, the value of the MLE point is 0.0.

boundary_col_indices: a vector that is empty for the BivariateConfidenceStruct type and of length two for the UnivariateConfidenceStruct type. Contains the column indices in points of the confidence interval parameters for the UnivariateConfidenceStruct type. Default is an empty vector.

PlaceholderLikelihood.AbstractConfidenceStruct — Type

AbstractConfidenceStruct

Supertype for confidence boundary storage structs.

Subtypes

UnivariateConfidenceStruct

BivariateConfidenceStruct

SampledConfidenceStruct

PlaceholderLikelihood.UnivariateConfidenceStruct — Type

UnivariateConfidenceStruct(confidence_interval::Vector{<:Float64}, 
    interval_points::PointsAndLogLikelihood)

Struct that stores the confidence interval of a given interest parameter as well as points sampled within (and outside) the confidence interval and their corresponding log-likelihood values.

Fields

confidence_interval: a vector of length two with the confidence interval for a given interest parameter. If an entry has value NaN, that side of the confidence interval is outside the corresponding bound on the interest parameter.

interval_points: a PointsAndLogLikelihood struct containing any points that have been evaluated inside or outside the interval by get_points_in_interval!, their corresponding log-likelihood function value and the column indices of the confidence_interval points in interval_points.points. Points can be evaluated and stored that are outside the confidence interval so that log-likelihood profile plots are defined outside of the confidence interval. interval_points.points is stored in column-wise order of increasing interest parameter magnitude.

Supertype Hiearachy

UnivariateConfidenceStruct <: AbstractConfidenceStruct <: Any

PlaceholderLikelihood.BivariateConfidenceStruct — Type

BivariateConfidenceStruct(confidence_boundary::Matrix{Float64}, 
    internal_points::PointsAndLogLikelihood = PointsAndLogLikelihood(zeros(size(x, 1), 0), zeros(0)))

Struct that stores samples produced by an AbstractBivariateMethod that are on the bivariate confidence boundary of two interest parameters at a given confidence level and, if save_internal_points=true, any internal points found during the method with their corresponding log-likelihood values. Use bivariate_methods() for a list of available methods (see bivariate_methods).

Fields

confidence_boundary: an array of boundary points stored in columns, with each row corresponding to the respective index of each model parameter. This array can contain points that are inside the bivariate confidence boundary if the method being used brackets between an internal point and a point on the user-provided bounds: these points will be on a user-provided parameter bound.

internal_points: a PointsAndLogLikelihood struct containing points and their corresponding log-likelihood values that were found during a method, if save_internal_points=true. Default is an empty PointsAndLogLikelihood struct (used if save_internal_points=false).

Supertype Hiearachy

BivariateConfidenceStruct <: AbstractConfidenceStruct <: Any

PlaceholderLikelihood.SampledConfidenceStruct — Type

SampledConfidenceStruct(points::Array{Float64}, 
    ll::Vector{<:Float64})

Struct that stores samples produced by an AbstractSampleType that are within the confidence boundary of sample_dimension interest parameters at a given confidence level, with their corresponding log-likelihood values.

Fields

points: an array of points stored in columns, with each row corresponding to the respective index of each model parameter.

ll: a vector of log-likelihood function values corresponding to the point in each column of points. This number is standardised so that regardless of whether the true log-likelihood function or an ellipse approximation of the function is evaluated, the value of the MLE point is 0.0.

Supertype Hiearachy

SampledConfidenceStruct <: AbstractConfidenceStruct <: Any

Profile Types

Profile type is a Struct that specifies whether the profile to be taken uses the true loglikelihood function or an ellipse approximation of the loglikelihood function centred at the MLE (with optional use of parameter bounds).

PlaceholderLikelihood.AbstractProfileType — Type

AbstractProfileType

Supertype for profile types.

Subtypes

LogLikelihood

AbstractEllipseProfileType

PlaceholderLikelihood.AbstractEllipseProfileType — Type

AbstractProfileType

Supertype for ellipse approximation profile types.

Subtypes

EllipseApprox

EllipseApproxAnalytical

Supertype Hiearachy

AbstractProfileType <: AbstractProfileType <: Any

PlaceholderLikelihood.LogLikelihood — Type

LogLikelihood()

Use the true log-likelihood function for confidence profile evaluation. The methods IterativeBoundaryMethod and RadialRandomMethod are recommended for use with this profile type.

Supertype Hiearachy

LogLikelihood <: AbstractProfileType <: Any

PlaceholderLikelihood.EllipseApprox — Type

EllipseApprox()

Use an ellipse approximation of the log-likelihood function centred at the MLE with use of parameter bounds for confidence profile evaluation. The method RadialMLEMethod is recommended for use with this profile type.

Supertype Hiearachy

EllipseApprox <: AbstractEllipseProfileType <: AbstractProfileType <: Any

PlaceholderLikelihood.EllipseApproxAnalytical — Type

EllipseApproxAnalytical()

Use an ellipse approximation of the log-likelihood function centred at the MLE without use of parameter bounds for confidence profile evaluation. As no parameter bounds are involved, it can be analytically evaluated. The method AnalyticalEllipseMethod is recommended for use with this profile type - it analytically samples points on the confidence profile boundary using EllipseSampling.jl. Other methods can be used, but they will all be unable to find interest parameter points outside user-provided parameter bounds (although nuisance parameters will be allowed outside these bounds).

Supertype Hiearachy

EllipseApproxAnalytical <: AbstractEllipseProfileType <: AbstractProfileType <: Any