Univariate Profiles

PlaceholderLikelihood.get_points_in_interval!

PlaceholderLikelihood.get_uni_confidence_interval

PlaceholderLikelihood.get_uni_confidence_interval_points

PlaceholderLikelihood.univariate_confidenceintervals!

PlaceholderLikelihood.univariate_confidenceintervals! — Function

univariate_confidenceintervals!(model::LikelihoodModel, 
    θs_to_profile::Vector{<:Int64}=collect(1:model.core.num_pars); 
    <keyword arguments>)

Computes likelihood-based confidence interval profiles for the provided θs_to_profile interest parameters, where θs_to_profile is a vector of Int corresponding to the parameter indexes in model.core.θnames. Saves these profiles by modifying model in place.

Arguments

model: a LikelihoodModel containing model information, saved profiles and predictions.

θs_to_profile: vector of parameters to profile, as a vector of model parameter indexes. Default is collect(1:model.core.num_pars), or all parameters.

Keyword Arguments

confidence_level: a number ∈ (0.0, 1.0) for the confidence level to evaluate the confidence interval. Default is 0.95 (95%).

profile_type: whether to use the true log-likelihood function or an ellipse approximation of the log-likelihood function centred at the MLE (with optional use of parameter bounds). Available profile types are LogLikelihood, EllipseApprox and EllipseApproxAnalytical. Default is LogLikelihood() (LogLikelihood).

use_existing_profiles: boolean variable specifying whether to use existing profiles of a parameter θi to decrease the width of the bracket used to search for the desired confidence interval using PlaceholderLikelihood.get_interval_brackets. Default is false.

θs_is_unique: boolean variable specifying whether all parameter indexes in θs_to_profile are ordered by parameter index (ascending) and unique. Default is false.

num_points_in_interval: an integer number of points to optionally evaluate within the confidence interval for each interest parameter using get_points_in_interval!. Points are linearly spaced in the interval and have their optimised log-likelihood value recorded. Useful for plots that visualise the confidence interval or for predictions from univariate profiles. Default is 0.

additional_width: a Real number greater than or equal to zero. Specifies the additional width to optionally evaluate outside the confidence interval's width if num_points_in_interval is greater than 0 using get_points_in_interval!. Half of this additional width will be placed on either side of the confidence interval. If the additional width goes outside a bound on the parameter, only up to the bound will be considered. The spacing of points in the additional width will try to match the spacing of points evaluated inside the interval. Useful for plots that visualise the confidence interval as it shows the trend of the log-likelihood profile outside the interval range. Default is 0.0.

existing_profiles: Symbol ∈ [:ignore, :overwrite] specifying what to do if profiles already exist for a given interest parameter, confidence_level and profile_type. See below for each symbol's meanings. Default is :merge.

show_progress: boolean variable specifying whether to display progress bars on the percentage of θs_to_profile completed and estimated time of completion. Default is model.show_progress.

use_distributed: boolean variable specifying whether to use a normal for loop or a @distributed for loop across combinations of interest parameters. The variable makes no difference if Distributed.jl is not being used - it's intended for use when simulating parameter confidence interval coverage (see check_univariate_parameter_coverage).

existing_profiles meanings

:ignore means profiles that already exist will not be recomputed.

:overwrite means profiles that already exist will be overwritten.

Details

Using univariate_confidenceinterval this function uses a bracketing method for each interest parameter in θs_to_profile (depending on the setting for existing_profiles if these profiles already exist). Updates model.uni_profiles_df for each successful profile and saves their results as a UnivariateConfidenceStruct in model.uni_profiles_dict, where the keys for the dictionary is the row number in model.uni_profiles_df of the corresponding profile.

Iteration Speed Of the Progress Meter

An iteration within the progress meter is specified as one iteration per side of the confidence interval found and an additional iteration for once points within the interval have been found if num_points_in_interval > 0. This means on a per interest parameter basis, there are either two or three iterations counted in time/it calculation.

univariate_confidenceintervals!(model::LikelihoodModel, 
    θs_to_profile::Vector{<:Symbol}; 
    <keyword arguments>)

Profiles only the provided θs_to_profile interest parameters, where θs_to_profile is a vector of Symbol corresponding to the parameter symbols in model.core.θnames.

univariate_confidenceintervals!(model::LikelihoodModel, 
    profile_m_random_combinations::Int; 
    <keyword arguments>)

Profiles m random interest parameters (sampling without replacement), where 0 < m ≤ model.core.num_pars.

θs_is_unique keyword argument

θs_is_unique is not a valid keyword argument for this function method as it internally produces the parameter combinations which are guaranteed to be unique.

PlaceholderLikelihood.get_points_in_interval! — Function

get_points_in_interval!(model::LikelihoodModel, 
    num_points_in_interval::Int; 
    confidence_levels::Vector{<:Float64}=Float64[], 
    profile_types::Vector{<:AbstractProfileType}=AbstractProfileType[], 
    additional_width::Real=0.0)

Evaluate and save num_points_in_interval linearly spaced points between the confidence intervals of existing univariate profiles that meet the requirements of PlaceholderLikelihood.desired_df_subset (see Keyword Arguments), as well as any additional width on the sides of the interval. Modifies model in place.

Arguments

model: a LikelihoodModel containing model information, saved profiles and predictions.

num_points_in_interval: an integer number of points to evaluate within the confidence interval. Points are linearly spaced in the interval and have their optimised log-likelihood value recorded. Useful for plots that visualise the confidence interval or for predictions from univariate profiles.

Keyword Arguments

confidence_levels: a vector of confidence levels or a Float64 of a single confidence level. If empty, all confidence levels of univariate profiles will be considered for finding interval points. Otherwise, only confidence levels of univariate profiles in confidence_levels will be considered.

profile_types: a vector of AbstractProfileType structs. If empty, all profile types of univariate profiles are considered. Otherwise, only univariate profiles with matching profile types will be considered.

additional_width: a Real number greater than or equal to zero. Specifies the additional width to optionally evaluate outside the confidence interval's width. Half of this additional width will be placed on either side of the confidence interval. If the additional width goes outside a bound on the parameter, only up to the bound will be considered. The spacing of points in the additional width will try to match the spacing of points evaluated inside the interval. Useful for plots that visualise the confidence interval as it shows the trend of the log-likelihood profile outside the interval range. Default is 0.0.

Details

Interval points and their corresponding log-likelihood values are stored in the interval_points field of a UnivariateConfidenceStruct. These are updated using PlaceholderLikelihood.update_uni_dict_internal!.

PlaceholderLikelihood.get_uni_confidence_interval — Function

get_uni_confidence_interval(model::LikelihoodModel, uni_row_number::Int)

Returns the confidence interval corresponding to the profile in row uni_row_number of model.uni_profiles_df as a vector of length two. If an entry has value NaN, that side of the confidence interval is outside the corresponding bound on the interest parameter.

PlaceholderLikelihood.get_uni_confidence_interval_points — Function

get_uni_confidence_interval_points(model::LikelihoodModel, uni_row_number::Int)

Returns the interval points PointsAndLogLikelihood struct corresponding to the profile in row uni_row_number of model.uni_profiles_df.